==> scripts/Physiology/2.MEM_field_flam_PV.Rmd <==
---
title: "Mixed Effects Models - Field Data, Flammability, PV Curves"
author: "Indra Boving"
date: "11/8/2022"
output: html_document
---

Thid code is to visualize and investigate the relationships between water potential and lfm in PV curves, Flammability benchtop drydowns, and in field-collected data.

# Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#library(datapasta)
library(ggplot2)
library(purrr) #for visualizing all variables in one big plot
library(naniar) #for dealing with NAs nicely 
library(tidyverse)
library(ggpubr)
library(cowplot)
#library(GGally) #for corr plots
#require(kimisc) # has the nlist function to create a named list
#library(strengejacke)
library(sjPlot) # table functions
library(here)
#library(effects) #for plofhng model effects
#library(sjstats) #use for r2 functions
library("broom.mixed")
library(MuMIn)
library(modelsummary)
library(nlme)
library(lme4)
library(MetBrewer)
#library(remef)
library(kableExtra)
library(janitor)
library(lubridate)
library(specr)
#library(cars)
#install.packages("specr")

filter = dplyr::filter
mutate = dplyr::mutate
select = dplyr::select
here = here::here

#source(here::here("scripts", "scripts_functions", "figure_info_sierra_flammability.R")) #color and theme info is here
```

# 1. PV Curves

```{r,warning=FALSE}
pv_all_df <- read_csv(here("processed-data", "all_pv_curves_clean.csv"), show_col_types = FALSE) %>% 
  filter(!spp %in% c("ADFA", "CEME")) %>% 
  mutate(sample = as.factor(sample)) %>% 
  mutate(month = as.factor(month)) %>% 
  mutate(date = ymd(date)) %>% 
  unite(individual, c('spp','sample','date'), sep = '_', remove = F) %>% 
  mutate(year_month = str_c(year, '_', month), 
         year = as.factor(year)) %>% 
  clean_names() %>% 
  group_by(spp) %>% 
  mutate(plant = udpipe::unique_identifier(individual), 
         plant = as.factor(plant),
         fresh_weight_saturated = as.numeric(fresh_weight_saturated)) %>% 
  group_by(spp, plant) %>% 
  mutate(maxLFM = ((fresh_weight_saturated - dry_weight)/dry_weight)*100) %>% 
   mutate(spp = fct_relevel(spp, "ABCO","PIJE","CADE","CECO", "ARPA","QUKE")) %>% 
  mutate(strategy = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "Isohydric", 
    spp == "PIJE" ~ "Isohydric", 
     spp == "CADE" ~ "Isohydric", 
    spp == "CECO" ~ "Anisohydric", 
    spp == "ARPA" ~ "Anisohydric", 
    spp == "QUKE" ~ "Anisohydric"
  )) %>% 
   mutate(strategy_binary = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "1", 
    spp == "ARPA" ~ "2", 
    spp == "CADE" ~ "1", 
    spp == "CECO" ~ "2", 
    spp == "PIJE" ~ "1", 
    spp == "QUKE" ~ "2"
  )) %>% 
   mutate(Species = case_when(
     spp == "ABCO" ~ "A. concolor",
    spp == "PIJE" ~ "P. jeffreyi", 
    spp == "CADE" ~ "C. decurrens", 
    spp == "CECO" ~ "C. cordulatus", 
    spp == "ARPA" ~ "A. patula", 
    spp == "QUKE" ~ "Q. kelloggii"
    )) %>% 
  mutate(type = case_when(
    spp == "PIJE" ~ "tree", 
    spp == "QUKE" ~ "tree", 
    spp == "CECO" ~ "shrub", 
    spp == "ARPA" ~ "shrub", 
    spp == "ABCO" ~ "tree", 
    spp == "CADE" ~ "tree"
  )) %>% 
  mutate(F.Group = case_when(
    spp == "ARPA" ~ "Angiosperm", 
    spp == "ABCO" ~ "Gymnosperm",
    spp == "CADE" ~ "Gymnosperm",
    spp == "CECO" ~ "Angiosperm",
    spp == "PIJE" ~ "Gymnosperm",
    spp == "QUKE" ~ "Angiosperm"
  )) %>% 
  filter(timing == "fall")%>% 
  mutate(ilfm = 1/lfm) %>% 
  filter(individual != 'QUKE_2_2020-10-21',
         individual !='QUKE_4_2020-10-21')

# model criticism plots
qqnorm(pv_all_df$lfm) #normally distributed
qqline(pv_all_df$lfm)

qqnorm(pv_all_df$ilfm)
qqline(pv_all_df$ilfm) #NOT normally distributed

write_csv(pv_all_df, (here("processed-data", "var-decomp", "all_pv_curves_clean_vardecomp.csv")))
```

## Summaries
```{r}
pv_summaries_2 <- read_csv(here("raw-data", "PV_sierra_summaries.csv")) %>% 
  clean_names() %>% 
   mutate(spp = fct_relevel(species, "ABCO","PIJE","CADE","CECO", "ARPA","QUKE"), 
          tlp = ytlp_m_pa, 
          lfmtlp = lf_mtlp)  %>% 
  mutate(strategy = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "Isohydric", 
    spp == "PIJE" ~ "Isohydric", 
     spp == "CADE" ~ "Isohydric", 
    spp == "CECO" ~ "Anisohydric", 
    spp == "ARPA" ~ "Anisohydric", 
    spp == "QUKE" ~ "Anisohydric"
  )) %>% 
   mutate(strategy_binary = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "1", 
    spp == "ARPA" ~ "2", 
    spp == "CADE" ~ "1", 
    spp == "CECO" ~ "2", 
    spp == "PIJE" ~ "1", 
    spp == "QUKE" ~ "2"
  )) %>% 
   mutate(spp = fct_relevel(spp, "ABCO","PIJE","CADE","CECO", "ARPA","QUKE"))  %>% 
   mutate(Species = case_when(
     spp == "ABCO" ~ "A. concolor",
    spp == "PIJE" ~ "P. jeffreyi", 
    spp == "CADE" ~ "C. decurrens", 
    spp == "CECO" ~ "C. cordulatus", 
    spp == "ARPA" ~ "A. patula", 
    spp == "QUKE" ~ "Q. kelloggii"
    )) %>% 
  mutate(type = case_when(
    spp == "PIJE" ~ "tree", 
    spp == "QUKE" ~ "tree", 
    spp == "CECO" ~ "shrub", 
    spp == "ARPA" ~ "shrub", 
    spp == "ABCO" ~ "tree", 
    spp == "CADE" ~ "tree"
  )) %>% 
  mutate(F.Group = case_when(
    spp == "ARPA" ~ "Angiosperm", 
    spp == "ABCO" ~ "Gymnosperm",
    spp == "CADE" ~ "Gymnosperm",
    spp == "CECO" ~ "Angiosperm",
    spp == "PIJE" ~ "Gymnosperm",
    spp == "QUKE" ~ "Angiosperm"
  )) %>% 
  mutate(date = lubridate::mdy(date), 
         year = year(date))

fgroup_pv_summaries <- pv_summaries_2 %>% 
  group_by(F.Group) %>% 
  mutate(mean_tlp_group = mean(tlp))
```
## Plot

```{r}
dat_text <- data.frame(
  F.Group = c("Angiosperm", "Gymnosperm"),
  label   = c(NA, "b")
)

pv_plot <- pv_all_df %>%
  filter(!spp %in% c("ADFA", "CEME")) %>%
  ggplot(aes(
    x = -1 * water_potential,
    y = lfm,
    color = spp,
   # shape = as.factor(year)
  )) +
  geom_vline(aes(xintercept = -1*mean_tlp_group, 
                 lty = F.Group), data = fgroup_pv_summaries) +
  geom_point(alpha = .5, 
              aes(shape = as.factor(year))
             ) +
  color_many2 +
  theme(legend.position = "none",
        axis.title.x = element_text("")) +
  labs(title = "PV Curves",
       x = "",
       y = "LFM") +
  xlim(0, 10) +
  # annotate(
  #   geom = "text",
  #   x = 9,
  #   y = 232,
  #   label = 'b',
  #   fontface = 'bold',
  #   size = 10
  # ) +  
  geom_text(
  data    = dat_text,
  color = "black",
  fontface = 'bold', size = 10,
  mapping = aes(x = 9, y = 232, label = label))+
  facet_wrap(~F.Group)+
  theme(plot.margin = unit(c(.2, .5,-.5, .25), "cm")) #margin(t = 0, r = 0, b = 0, l = 0, unit = "pt")
  
pv_plot
```

F.test: 
Allowa us to compare the variance between two groups, in this case angiosperms and gymnosperms:
```{r}
res.ftest <- var.test(lfm ~ F.Group, data = pv_all_df)
res.ftest
#There is a significant difference between the variaance 
```

Var decomp: 
https://cran.r-project.org/web/packages/specr/vignettes/decompose_var.html 
```{r}
pv_mod <- lmer(lfm~(1|water_potential) + (1|F.Group) + (1|plant), data = pv_all_df)
summary(pv_mod)
anova(pv_mod)

specr::icc_specs(pv_mod) %>%
  mutate_if(is.numeric, round, 2)
```


## Plot (-MPa)
```{r}
dat_text <- data.frame(
  F.Group = c("Angiosperm", "Gymnosperm"),
  label   = c(NA, "b")
)

pv_plot_neg <- pv_all_df %>%
  filter(!spp %in% c("ADFA", "CEME")) %>%
  ggplot(aes(
    x = -1/water_potential,
    y = lfm,
    color = spp
  )) +
  geom_vline(aes(xintercept = -1/mean_tlp_group, 
                 lty = F.Group), data = fgroup_pv_summaries) +
  geom_point(alpha = .5) +
  color_many2 +
  theme(legend.position = "none",
        axis.title.x = element_text("")) +
  labs(title = "PV Curves",
       x = "",
       y = "LFM") +
  xlim(0, 15) +
  # annotate(
  #   geom = "text",
  #   x = 9,
  #   y = 232,
  #   label = 'b',
  #   fontface = 'bold',
  #   size = 10
  # ) +  
  geom_text(
  data    = dat_text,
  color = "black",
  fontface = 'bold', size = 10,
  mapping = aes(x = 14, y = 232, label = label))+
  facet_wrap(~F.Group)+
  theme(plot.margin = unit(c(.2, .5,-.5, .25), "cm")) #margin(t = 0, r = 0, b = 0, l = 0, unit = "pt")
  
pv_plot_neg
```

## MEM
```{r}
m10.5 <- lmer(lfm~ water_potential*F.Group + (1|plant), data = pv_all_df)
summary(m10.5)
AIC(m10.5)

m10 <- lmer(lfm~ water_potential + F.Group + (1|plant), data = pv_all_df)
summary(m10)
AIC(m10)

anova(m10, m10.5) #don't need interaction

m10.75 <- lmer(lfm~ water_potential + F.Group + as.factor(year) + (1|plant), data = pv_all_df)
summary(m10.75) #yes add year

anova(m10, m10.75)

m10.757 <- lmer(lfm~ water_potential*as.factor(year) + F.Group + (1|plant), data = pv_all_df)
summary(m10.757)

anova(m10.75, m10.757) # dont need year interaction

qqnorm(resid(m10.75))
r <- resid(m10.75)
plot(fitted(m10.75), r)


#qqnorm(m10.75, ~ranef(., level=1))

anova(mm10.75anova(m10.75))
car::Anova(m10.75)

specr::icc_specs(m10.75) %>%
  mutate_if(is.numeric, round, 2)
```

# 2. Flam. Drydown

```{r, warning=FALSE}
flam_curve_phys_all_raw <- read_csv(here("processed-data", "sierra_flam_data_all.csv"), show_col_types = FALSE)

flam_curve_df <- flam_curve_phys_all_raw %>% 
   filter(!spp %in% c("ADFA", "CEME")) %>% 
  mutate(water_potential = -1*pos_mpa, 
         swc = swc_sat, 
         month = as.factor(month)) %>% 
  group_by(spp) %>% 
  mutate(plant = udpipe::unique_identifier(individual), 
         maxLFM = gww_gdw_saturated * 100) %>% 
  mutate(spp = toupper(spp)) %>% 
  mutate(strategy = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "Isohydric", 
    spp == "PIJE" ~ "Isohydric", 
     spp == "CADE" ~ "Isohydric", 
    spp == "CECO" ~ "Anisohydric", 
    spp == "ARPA" ~ "Anisohydric", 
    spp == "QUKE" ~ "Anisohydric"
  )) %>% 
   mutate(strategy_binary = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "1", 
    spp == "ARPA" ~ "2", 
    spp == "CADE" ~ "1", 
    spp == "CECO" ~ "2", 
    spp == "PIJE" ~ "1", 
    spp == "QUKE" ~ "2"
  )) %>% 
   mutate(Species = case_when(
     spp == "ABCO" ~ "A. concolor",
    spp == "PIJE" ~ "P. jeffreyi", 
    spp == "CADE" ~ "C. decurrens", 
    spp == "CECO" ~ "C. cordulatus", 
    spp == "ARPA" ~ "A. patula", 
    spp == "QUKE" ~ "Q. kelloggii"
    )) %>% 
  mutate(type = case_when(
    spp == "PIJE" ~ "tree", 
    spp == "QUKE" ~ "tree", 
    spp == "CECO" ~ "shrub", 
    spp == "ARPA" ~ "shrub", 
    spp == "ABCO" ~ "tree", 
    spp == "CADE" ~ "tree"
  )) %>% 
  mutate(F.Group = case_when(
    spp == "ARPA" ~ "Angiosperm", 
    spp == "ABCO" ~ "Gymnosperm",
    spp == "CADE" ~ "Gymnosperm",
    spp == "CECO" ~ "Angiosperm",
    spp == "PIJE" ~ "Gymnosperm",
    spp == "QUKE" ~ "Angiosperm"
  )) %>% 
  select(lfm, water_potential, mpa, year, month, year_month, plant, F.Group, spp) %>% 
  mutate(ilfm = 1/lfm)

# model criticism plots
qqnorm(flam_curve_df$lfm) #normally distributed
qqline(flam_curve_df$lfm)

qqnorm(flam_curve_df$ilfm)
qqline(flam_curve_df$ilfm) #NOT normally distributed

write_csv(flam_curve_df, here("processed-data", "var-decomp", "flam_curve_vardecomp.csv"))
```


## Plot
```{r}
dat_text <- data.frame(
  F.Group = c("Angiosperm", "Gymnosperm"),
  label   = c(NA, "c")
)

flam_plot <- flam_curve_df %>% 
  filter(lfm < 300, 
         lfm > 50) %>% 
  filter(!spp %in% c("ADFA", "CEME")) %>% 
  ggplot(aes(y = lfm, 
             x = -1*water_potential, 
             color = spp)) +
  geom_vline(aes(xintercept = -1*mean_tlp_group, 
                 lty = F.Group), data = fgroup_pv_summaries) +
  geom_point(alpha = .5, 
             aes( shape = as.factor(year))
             ) +
 # geom_smooth(method = "lm", se = F) +
  color_many2 +
  theme(legend.position = "none")+
  labs(title = "Flam", 
       x = "Water Potential", 
       y = "LFM")+
  theme(plot.margin = unit(c(.2, .5, .3, .25), "cm")) +
 # xlim(0, 9) + 
  geom_text(
  data    = dat_text,
  color = "black",
  fontface = 'bold', size = 10,
  mapping = aes(x = 9, y = 210, label = label))+
 # annotate(geom = "text", x = 9, y = 225, label = 'c', fontface = 'bold', size = 10)+
  facet_wrap(~F.Group)
flam_plot
```

## Plot (-MPa)
```{r}
dat_text <- data.frame(
  F.Group = c("Angiosperm", "Gymnosperm"),
  label   = c(NA, "c")
)

flam_plot_neg <- flam_curve_df %>% 
  filter(lfm < 300, 
         lfm > 50) %>% 
  filter(!spp %in% c("ADFA", "CEME")) %>% 
  ggplot(aes(y = lfm, 
             x = -1/water_potential, 
             color = spp)) +
  geom_vline(aes(xintercept = -1/mean_tlp_group, 
                 lty = F.Group), data = fgroup_pv_summaries) +
  geom_point(alpha = .5) +
 # geom_smooth(method = "lm", se = F) +
  color_many +
  theme(legend.position = "none")+
  labs(title = "Flam", 
       x = "Water Potential", 
       y = "LFM")+
  theme(plot.margin = unit(c(.2, .5, .3, .25), "cm")) +
  xlim(0, 15) + 
  geom_text(
  data    = dat_text,
  color = "black",
  fontface = 'bold', size = 10,
  mapping = aes(x = 14, y = 210, label = label))+
 # annotate(geom = "text", x = 9, y = 225, label = 'c', fontface = 'bold', size = 10)+
  facet_wrap(~F.Group)
flam_plot_neg
```

## MEM
```{r}
m11.5 <- lmer(lfm~ water_potential + F.Group + (1|plant), data = flam_curve_df)
summary(m11.5)

m11 <- lmer(lfm~ water_potential*F.Group + (1|plant), data = flam_curve_df)

anova(m11, m11.5) #Yes interaction between mpa and f group

m11.75  <- lmer(lfm~ water_potential*F.Group + as.factor(year) + (1|plant), data = flam_curve_df)
m11.75 #there shouldn't be an interaction between year and f group, so dont include
summary(m11.75)
anova(m11, m11.75) 

qqnorm(resid(m11.75)) #tells you if your residuals (points around the fitted line) are distributed the way your error distribution should be (answers: are your errors normally distributed). It will tell you about your outliers (points far from line). Systematic trends would show you that you're using the wrong error function in your model. 

r <- resid(m11.75)
plot(fitted(m11.75), r)

anova(m11.75)
car::Anova(m11.75)

performance::multicollinearity(m11.75)

specr::icc_specs(m11) %>%
  mutate_if(is.numeric, round, 2)

plot(m11.75)

#big cooks distance would mean there are outliers that are important. This is like the final outlier check. ("Leverage plot" to see if points have pull on the data).

#heterscedscity (homogeneous variance); want this to be relatively unbiased. In ecological data its common that its not fully evenly spread. Tell us if it's the right model type. Less important than fitting the right model parameter, because you'll still get the correct line, you'll just not get the correct error around your estimates.

#Also do this: 
#Check if random effects are normally distributed. QQplot for the random effects and residuals of the random effects category. (Asks: is there a site/tree/plot that acts really differently?)


```


# 3. Field Data
```{r}
field_data_all <- read_csv(here("processed-data", "field_data_2020_2021.csv"), show_col_types = FALSE) %>% 
    filter(lfm < 300, 
         lfm > 50, 
        # age == "new"
         ) %>% 
 # select(-mpa, -water_potential) %>% 
  mutate(water_potential = midday, 
         unique_id2 = str_c(spp, "_",pod, '_', site),
         plant = udpipe::unique_identifier(unique_id2)) %>% 
  mutate(year_month = str_c(year, '_', month)) %>% 
 # group_by(spp, plant) %>% 
  mutate(maxLFM = wettest_lfm)  %>% 
  mutate(strategy = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "Isohydric", 
    spp == "PIJE" ~ "Isohydric", 
     spp == "CADE" ~ "Isohydric", 
    spp == "CECO" ~ "Anisohydric", 
    spp == "ARPA" ~ "Anisohydric", 
    spp == "QUKE" ~ "Anisohydric"
  )) %>% 
   mutate(strategy_binary = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "1", 
    spp == "ARPA" ~ "2", 
    spp == "CADE" ~ "1", 
    spp == "CECO" ~ "2", 
    spp == "PIJE" ~ "1", 
    spp == "QUKE" ~ "2"
  )) %>% 
   mutate(Species = case_when(
     spp == "ABCO" ~ "A. concolor",
    spp == "PIJE" ~ "P. jeffreyi", 
    spp == "CADE" ~ "C. decurrens", 
    spp == "CECO" ~ "C. cordulatus", 
    spp == "ARPA" ~ "A. patula", 
    spp == "QUKE" ~ "Q. kelloggii"
    )) %>% 
  mutate(type = case_when(
    spp == "PIJE" ~ "tree", 
    spp == "QUKE" ~ "tree", 
    spp == "CECO" ~ "shrub", 
    spp == "ARPA" ~ "shrub", 
    spp == "ABCO" ~ "tree", 
    spp == "CADE" ~ "tree"
  )) %>% 
  mutate(F.Group = case_when(
    spp == "ARPA" ~ "Angiosperm", 
    spp == "ABCO" ~ "Gymnosperm",
    spp == "CADE" ~ "Gymnosperm",
    spp == "CECO" ~ "Angiosperm",
    spp == "PIJE" ~ "Gymnosperm",
    spp == "QUKE" ~ "Angiosperm"
  )) %>% 
  mutate(ilfm = 1/lfm) 

field_data_all_2 <- field_data_all %>% 
  drop_na(lfm, water_potential, plant, F.Group, year)

write_csv(field_data_all, (here("processed-data", "var-decomp", "field_data_vardecomp.csv")))

# model criticism plots
qqnorm(field_data_all$lfm)
qqline(field_data_all$lfm) #NOT Normal

qqnorm(field_data_all$ilfm)
qqline(field_data_all$ilfm) #Normal-ish
```
## Plot
```{r}
dat_text <- data.frame(
  F.Group = c("Angiosperm", "Gymnosperm"),
  label   = c(NA, "a")
)

field_plot <- field_data_all %>% 
  filter(lfm < 300, 
         lfm > 50) %>% 
  ggplot(aes(y = lfm, 
             x = -1*water_potential, 
             color = spp)) +
  geom_vline(aes(xintercept = -1*mean_tlp_group, 
                 lty = F.Group), data = fgroup_pv_summaries) +
  geom_point(alpha = .5, 
             aes(shape = as.factor(year))
             )+
  facet_wrap(~F.Group) +
 # geom_smooth(method = 'nls', formula = 'y~x*exp(1)') + 
  geom_smooth(method = "loess", se = F) +
  color_many2 +
  theme(legend.position = "none") +
  labs(title = "Field", 
       x = "", 
       y = "LFM") +
  theme(plot.margin = unit(c(.2, .5, -.5, .25), "cm")) +
  xlim(0, 10) +
  geom_text(
  data    = dat_text,
  color = "black",
  fontface = 'bold', size = 10,
  mapping = aes(x = 9, 
                y = 300, 
                #y = 2,
                label = label)
 )
  #annotate(geom = "text", x = 9, y = 300, label = 'a', fontface = 'bold', size = 10) 
field_plot
```
## ****Plot inverse LFM (1/lfm) (IN SI)
```{r}
field_plot_ilfm<- field_data_all %>% 
  filter(lfm < 300, 
         lfm > 50) %>% 
  ggplot(aes(y = 1/lfm, 
             x = water_potential, 
             color = spp)) +
  geom_point(alpha = .3, 
             #aes(shape = as.factor(year) )
             )+
  facet_wrap(~F.Group, scales = "free") +
  geom_smooth(method = "lm", se = F) +
  color_many +
  theme(legend.position = "none") +
  labs(title = "Field", 
       x = "Water Potential", 
       y = "1/LFM") +
  theme(plot.margin = unit(c(.2, .5, .5, .25), "cm")) 
field_plot_ilfm
```
## Plot (log-LFM)
```{r}
field_plot_logged<- field_data_all %>% 
  filter(lfm < 300, 
         lfm > 50) %>% 
  ggplot(aes(y = log(lfm), 
             x = -1*water_potential, 
             color = spp)) +
  geom_vline(aes(xintercept = -1*mean_tlp_group, 
                 lty = F.Group), data = fgroup_pv_summaries) +
  geom_point(alpha = .5, 
             aes(shape = as.factor(year))
             )+
  facet_wrap(~F.Group) +
 # geom_smooth(method = "lm", se = F) +
  color_many +
  theme(legend.position = "none") +
  labs(title = "Field", 
       x = "", 
       y = "LFM") +
  theme(plot.margin = unit(c(.2, .5, -.5, .25), "cm")) +
  xlim(0, 10) +
  geom_text(
  data    = dat_text,
  color = "black",
  fontface = 'bold', size = 10,
  mapping = aes(x = 9, 
               # y = 300, 
                y = 2,
                label = label)
 )
  #annotate(geom = "text", x = 9, y = 300, label = 'a', fontface = 'bold', size = 10) 
field_plot_logged
```

## Plot (-MPa)
```{r}
dat_text <- data.frame(
  F.Group = c("Angiosperm", "Gymnosperm"),
  label   = c(NA, "a")
)

field_plot_neg <- field_data_all %>% 
  filter(lfm < 300, 
         lfm > 50) %>% 
  ggplot(aes(y = lfm, 
             x = -1/water_potential, 
             color = spp)) +
  geom_vline(aes(xintercept = -1/mean_tlp_group, 
                 lty = F.Group), data = fgroup_pv_summaries) +
  geom_point(alpha = .5)+
  facet_wrap(~F.Group) +
 # geom_smooth(method = "lm", se = F) +
  color_many +
  theme(legend.position = "none") +
  labs(title = "Field", 
       x = "", 
       y = "LFM") +
  theme(plot.margin = unit(c(.2, .5, -.5, .25), "cm")) +
  xlim(0, 2) +
  geom_text(
  data    = dat_text,
  color = "black",
  fontface = 'bold', size = 10,
  mapping = aes(x = 1.7, y = 300, label = label))
  #annotate(geom = "text", x = 9, y = 300, label = 'a', fontface = 'bold', size = 10) 
field_plot_neg
```


We should visualize in LFM~MPa space for the sake of comparing to benchtop curves, but transforming to 1/LFM is actually normally distributed (could show this in SI?).


## MEM
```{r}
m12.5 <- lm(ilfm~ water_potential*F.Group + as.factor(year), data = field_data_all)
summary(m12.5)
plot(m12.5, which = 1) 
plot(m12.5, which = 2)
plot(m12.5, which = 5)


m12.5a <- lm(ilfm~ water_potential*F.Group*age + as.factor(year), data = field_data_all)
summary(m12.5a)
plot(m12.5a, which = 1) 
plot(m12.5a, which = 2)
plot(m12.5a, which = 5)

m12.5ab <- lm(ilfm~ water_potential*F.Group + age + as.factor(year), data = field_data_all)
summary(m12.5ab)  

AIC(m12.5, m12.5a,m12.5ab) #
```


```{r}
lmer12 <- lmer(lfm~ water_potential*F.Group + as.factor(year)+ (1|plant), data = field_data_all)
summary(lmer12)
plot(lmer12, which = 1) 
qqnorm(residuals(lmer12)) 
qqline(residuals(lmer12))

#checking model random effects: 
modrandomeffects <- ranef(lmer12)
qqnorm(modrandomeffects$plant$`(Intercept)`)
qqline(modrandomeffects$plant$`(Intercept)`)
library(car)
qqp(modrandomeffects$plant$`(Intercept)`)
modrandomeffects$plant

lmer12.5 <- lmer(ilfm~ water_potential*F.Group + as.factor(year)+ (1|plant), data = field_data_all)
summary(lmer12.5)
plot(lmer12.5, which = 1) 
qqnorm(residuals(lmer12.5)) 
qqline(residuals(lmer12.5))


summary(m12.5)

#Transformed model: 
#y = a/(b+x)
#nonlin12.4 <- nls(lfm~)
```

####Just Angiosperms: 

```{r}
angios_field_data <- field_data_all %>% 
  filter(F.Group == "Angiosperm")

hist(angios_field_data$lfm)
hist(angios_field_data$ilfm)

m12.5a <- lm(ilfm~ water_potential, data = angios_field_data )
summary(m12.5a)
plot(m12.5a, which = 1) 
plot(m12.5a, which = 2)
plot(m12.5a, which = 5)

lmer12a <- lmer(lfm~ water_potential + as.factor(year)+ (1|plant), data = angios_field_data )
summary(lmer12a)
plot(lmer12a, which = 1) 
qqnorm(residuals(lmer12a)) 
qqline(residuals(lmer12a))


lmer12.5a <- lmer(ilfm~ water_potential + as.factor(year)+ (1|plant), data = angios_field_data )
summary(lmer12.5a)
plot(lmer12.5a, which = 1) 
qqnorm(residuals(lmer12.5a)) 
qqline(residuals(lmer12.5a))
```

####Just Gymnosperms: 

```{r}
gymnos_field_data <- field_data_all %>% 
  filter(F.Group == "Gymnosperm")

hist(gymnos_field_data$lfm)
hist(gymnos_field_data$ilfm)

m12.5g <- lm(ilfm~ water_potential, data = gymnos_field_data )
summary(m12.5g)
plot(m12.5g, which = 1) 
plot(m12.5g, which = 2)
plot(m12.5g, which = 5)

lmer12g <- lmer(lfm~ water_potential + as.factor(year)+ (1|plant), data = angios_field_data )
summary(lmer12g)
plot(lmer12g, which = 1) 
qqnorm(residuals(lmer12g)) 
qqline(residuals(lmer12g))

lmer12g <- lmer(log(lfm)~ log(water_potential*-1) + as.factor(year)+ (1|plant), data = angios_field_data )
summary(lmer12g)
plot(lmer12g, which = 1) 
qqnorm(residuals(lmer12g)) 
qqline(residuals(lmer12g))


lmer12.5g <- lmer(ilfm~ water_potential + as.factor(year)+ (1|plant), data = angios_field_data )
summary(lmer12.5g)
plot(lmer12.5g, which = 1) 
qqnorm(residuals(lmer12.5g)) 
qqline(residuals(lmer12.5g))
```
# 4. Arranging/Combining

## Scatterplots
```{r, fig.height=6, fig.width=5}
field_plot_leg <- field_data_all %>% 
  filter(lfm < 300, 
         lfm > 50) %>% 
  ggplot(aes(y = lfm, 
             x = -1*water_potential, 
             color = spp)) +
  geom_vline(aes(xintercept = -1*mean_tlp_group, 
                 lty = F.Group), data = fgroup_pv_summaries) +
  geom_point(alpha = .5, 
             aes(shape = as.factor(year))) +
  labs(color = "Species", 
       lty = "TLP", 
       shape = "Year") +
 # geom_smooth(method = "lm", se = F) +
  color_many 
legend <- cowplot::get_legend(field_plot_leg)

#Mpas:
plots <- cowplot::plot_grid(field_plot, pv_plot, flam_plot,
          ncol = 1, 
         nrow = 3
          )
plots_legend <- cowplot::plot_grid(plots, legend, rel_widths = c(5, 1), rel_heights = c(-.1, -0.1, -.1))
plots_legend

ggsave(plot=plots_legend, bg = "white" , here("figures", 'extra-figures', "flam_pv_field"), device="jpg")
```

## -MPa Plots
```{r, fig.height=6, fig.width=5}
#Mpas:
plots_neg <- cowplot::plot_grid(field_plot_neg, pv_plot_neg, flam_plot_neg,
          ncol = 1, 
         nrow = 3
          )
plots_legend_neg <- cowplot::plot_grid(plots_neg, legend, rel_widths = c(5, 1), rel_heights = c(-.1, -0.1, -.1))
plots_legend_neg

ggsave(plot=plots_legend, bg = "white" , here::here("figures", "extra-figures", "flam_pv_field_neg"), device="jpg")
```

## MEM
```{r}
tab_model <- sjPlot::tab_model(m10.75, m11.75, m12.4,
                               show.reflvl = TRUE, 
          digits = 3, 
          show.aic = TRUE, 
          show.ci = FALSE,
          show.icc = FALSE, 
         string.pred = "Coeffcient", 
        dv.labels = c("PV", "Flam", "Field"),
         title = "LFM ~ MPa",
  string.p = "P-Value", 
  p.style = "stars")
tab_model
```

#5. For Pepperwood/to share

```{r}
pepperwood_df <- field_data_all_2 %>%
  select(spp, Species, unique_id, age_new, date, year, location, lfm, predawn, midday, location) %>%
   filter(lfm < 300,
         lfm > 50) %>%
  rename(tissue_age = age_new) %>% 
  pivot_longer(cols = c(predawn, midday),
               names_to = "time",
               values_to = "water_potential")

pepperwood_df %>%
  ggplot(aes(y = lfm,
             x = water_potential,
             color = as.factor(year),
             shape = tissue_age)) +
  geom_point() +
  facet_wrap(~Species)

pepperwood_df %>%
  mutate(month = month(date)) %>%
  filter(tissue_age == "old",
         year == 2021) %>%
  ggplot(aes(y = lfm,
             x = water_potential,
             color = as.factor(month),
            # shape = tissue_age
            # shape = as.factor(month))
         )) +
  geom_point(aes(shape = tissue_age)) +
  geom_smooth(method = "lm", se = F)

write.csv(pepperwood_df, here('processed-data', 'sierra_lfm_mpa_withpredawns.csv'))

write.csv(pepperwood_df %>% filter(time %in% c("midday")), here('processed-data', 'sierra_lfm_mpa.csv'))


```
