==> scripts/Physiology/4.2.phenology_and_LFM_boxplots_UPDATED.Rmd <==
---
title: "LFM Boxplots w/ Phenology Barplots"
author: "Indra Boving"
date: "1/29/2022"
output: html_document
---

This script contains the phenology x LFM boxplots figure as well as exploratory analyses necessary to make the figure. Two versions of this figure are in this script -- the first version uses unedited phenology data, while the second version uses edited phenology data informed by the average phenophases of each of the species as a population rather than focusing on individual plants. Also included in this script is the LFM vs. MPa figures and exploratory analyses looking at hydroscapes, thus investigating drought strategies of our species-of-interest.

# Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ggpubr)
library(tidyverse)
#install.packages("cowplot")
library(cowplot)
library(here)
library(lubridate)
month = lubridate::month
week = lubridate::week
library(janitor)
library(tidytext)
library(grid)
library(gridExtra)
library(ggbeeswarm)
library(wesanderson)
library(MetBrewer)
library(RColorBrewer)
select = dplyr::select
here = here::here

source(here("scripts", "scripts_functions", "figure_info_sierra_flammability.R")) #color and theme info is here
```

# Data 
This dataset has field-collected LFM and Mpa data, and LFM at TLP and MPa at TLP from PV Curves

### Field Data
From 0_water_content_all.Rmd
```{r}
field_data_all_3 <- read_csv(here("processed-data","field_data_2020_2021.csv"), show_col_types = FALSE) %>% 
  drop_na(spp) %>% 
  filter(lfm < 350)
```

### Phenology Data
Cleaning up our data so it is easier to use in this script
```{r}
pheno_df_raw <- read.csv(here("raw-data", "phenology_czo_2021.csv")) %>% 
  clean_names() %>% 
  mutate(date = ymd(observation_date),
         year = year(date), 
         month = month(date)) %>% 
  mutate(class = case_when(
    common_name == "incense cedar" ~ "Gymnosperm", 
    common_name == "white fir" ~ "Gymnosperm", 
    common_name == "Jeffrey pine" ~ "Gymnosperm", 
    common_name == "greenleaf manzanita" ~ "Angiosperm", 
    common_name == "whitethorn ceanothus" ~ "Angiosperm", 
    common_name == "California black oak" ~ "Angiosperm"
  )) %>% 
  mutate(code = case_when(
    common_name == "incense cedar" ~ "CADE", 
    common_name == "white fir" ~ "ABCO", 
    common_name == "Jeffrey pine" ~ "PIJE", 
    common_name == "greenleaf manzanita" ~ "ARPA", 
    common_name == "whitethorn ceanothus" ~ "CECO", 
    common_name == "California black oak" ~ "QUKE"
  )) %>% 
  group_by(phenophase_name, code) %>% 
  mutate(start_phase = min(date), 
         end_phase = max(date))

pheno_df <- pheno_df_raw %>% 
  select(-observation_id, -observation_time, -observation_date, -raw_abundance_value) %>% 
   mutate(spp = fct_relevel(code, "ABCO","PIJE","CADE","ARPA", "CECO", "QUKE"))  

npn_df <- read_csv(here("processed-data", "npn_data.csv")) # National phenology network dataset
```

### PV Summaries
```{r}
pv_summaries <- read_csv(here("raw-data", "PV_sierra_summaries_fixed.csv")) %>% 
  clean_names() %>% 
   mutate(spp = fct_relevel(species, "ABCO","PIJE","CADE","CECO", "ARPA","QUKE"))  %>% 
  mutate(strategy = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "Isohydric", 
    spp == "PIJE" ~ "Isohydric", 
     spp == "CADE" ~ "Isohydric", 
    spp == "CECO" ~ "Anisohydric", 
    spp == "ARPA" ~ "Anisohydric", 
    spp == "QUKE" ~ "Anisohydric"
  )) %>% 
   mutate(strategy_binary = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "1", 
    spp == "ARPA" ~ "2", 
    spp == "CADE" ~ "1", 
    spp == "CECO" ~ "2", 
    spp == "PIJE" ~ "1", 
    spp == "QUKE" ~ "2"
  )) %>% 
   mutate(spp = fct_relevel(spp, "ABCO","PIJE","CADE","CECO", "ARPA","QUKE"))  %>% 
   mutate(Species = case_when(
    spp == "ABCO" ~ "Ab. concolor",
    spp == "PIJE" ~ "Pi. jeffreyi", 
    spp == "CADE" ~ "Ca. decurrens", 
    spp == "CECO" ~ "Ce. cordulatus", 
    spp == "ARPA" ~ "Ar. patula", 
    spp == "QUKE" ~ "Qu. kelloggii"
    )) %>% 
  mutate(type = case_when(
    spp == "PIJE" ~ "tree", 
    spp == "QUKE" ~ "tree", 
    spp == "CECO" ~ "shrub", 
    spp == "ARPA" ~ "shrub", 
    spp == "ABCO" ~ "tree", 
    spp == "CADE" ~ "tree"
  )) %>% 
  mutate(F.Group = case_when(
    spp == "ARPA" ~ "Angiosperm", 
    spp == "ABCO" ~ "Gymnosperm",
    spp == "CADE" ~ "Gymnosperm",
    spp == "CECO" ~ "Angiosperm",
    spp == "PIJE" ~ "Gymnosperm",
    spp == "QUKE" ~ "Angiosperm"
  )) %>% 
  mutate(date = lubridate::mdy(date), 
         year = year(date))
```
#------------------------
## FIG. 5: Species, Age 
```{r}
field_data_all_3 <- field_data_all_3 %>% 
  mutate(Age = case_when(
    age == 'new' ~ 'New',
    age == 'old' ~ 'Old'
  )) %>% 
  mutate(midday = -1*midday)

lfm_vs_mpa_tissue_status_plot <- field_data_all_3 %>%
  #filter(month < 10) %>% #removing post-rain values (October sample collection)
  mutate(spp = fct_relevel(spp, "ABCO", "PIJE", "CADE", "ARPA", "QUKE", "CECO"))  %>%
  mutate(month = month(date, label = TRUE, abbr = TRUE)) %>%
  mutate(midday = -1*midday) %>% 
  mutate(Age = case_when(
    Age == "New" ~ "New Growth",
    Age == "Old" ~ "Old Growth"
  )) %>% 
  ggplot(aes(
    x = midday,
    y = lfm,
    #shape = Age
  )) +
  geom_point(size = 1, alpha = .65, 
             aes(color = Age)) +
  geom_hline(
    aes(yintercept = lfm_tlp),
    size = 0.7,
    #color = color_many2,
    alpha = .6,
    show_guide = F
  ) +
  geom_vline(
    aes(xintercept = -1*tlp, #lty = Species
        ),
    size = 0.7,
    color = 'black',
    alpha = .6,
    show_guide = F,
    lty = "dashed"
  ) +
  xlim(0, 6) +
  labs(
    y = "Live Fuel Moisture",
    x = "Midday Water Potential (MPa)",
    color = "Tissue Status"
  ) +
  ylim(50, 300) +
  facet_wrap( ~ spp, labeller = labeller(spp = spp.labs)) +
  scale_color_manual(values = c("#b3beb6", "#262E43")) +
  theme(
    strip.text = element_text(size = 12, face = "italic"),
    axis.title = element_text(size = 16, face = "bold"),
    legend.title = element_text(face = "bold"),
    legend.text = element_text(size  = 12),
    title = element_text(size = 16, face = 'bold', hjust = 0.5), 
    strip.background = element_rect(fill = "gray90"),
    legend.key = element_blank())
  
#scale_color_gradientn(colors=met.brewer("Morgenstern"))
lfm_vs_mpa_tissue_status_plot

ggsave(here("figures", "main-figures", "FigS1.LFM_vs_Midday_spp.jpg"),
       height = 5, width = 8,
       plot = lfm_vs_mpa_tissue_status_plot)
```

## FIG. 5: Species, Age 
```{r}
field_data_all_3 <- field_data_all_3 %>% 
  mutate(Age = case_when(
    age == 'new' ~ 'New',
    age == 'old' ~ 'Old'
  )) %>% 
  mutate(midday = -1*midday)

lfm_vs_mpa_tissue_status_plot <- field_data_all_3 %>%
  #filter(month < 10) %>% #removing post-rain values (October sample collection)
  mutate(spp = fct_relevel(spp, "ABCO", "PIJE", "CADE", "ARPA", "QUKE", "CECO"))  %>%
  mutate(month = month(date, label = TRUE, abbr = TRUE)) %>%
  mutate(midday = -1*midday) %>% 
  mutate(Age = case_when(
    Age == "New" ~ "New",
    Age == "Old" ~ "Old"
  )) %>% 
  ggplot(aes(
    x = midday,
    y = lfm,
    #shape = Age
  )) +
  geom_point(size = 1, alpha = .65, 
             aes(color = Age)) +
  geom_hline(
    aes(yintercept = lfm_tlp),
    size = 0.7,
    #color = color_many2,
    alpha = .6,
    show_legend = F
  ) +
 geom_vline(
    aes(xintercept = tlp),
    linewidtch        = 0.7,
    color       = "black",
    alpha       = 0.6,
    linetype    = "dashed",
    show.legend = FALSE
  ) +
  xlim(-6, 0) +
  labs(
    y = "Live Fuel Moisture",
    x = "Midday Water Potential (MPa)",
    color = "Tissue Status"
  ) +
  ylim(50, 300) +
  facet_wrap( ~ spp, 
              labeller = labeller(spp = spp.labs)) +
  scale_color_manual(values = c("#b3beb6", "#262E43")) +
  theme(
    strip.text      = element_text(size = 12, face = "italic"),
    axis.title      = element_text(size = 16, face = "bold"),
    legend.title    = element_text(face = "bold"),
    legend.text     = element_text(size = 12),
    title           = element_text(size = 16, face = "bold", hjust = 0.5),
    strip.background = element_rect(fill = "gray90"),
    legend.key      = element_blank()
  )
  
#scale_color_gradientn(colors=met.brewer("Morgenstern"))
lfm_vs_mpa_tissue_status_plot

ggsave(here("figures", "main-figures", "FigS1.LFM_vs_Midday_spp_reversex.jpg"),
       height = 5, width = 8,
       plot = lfm_vs_mpa_tissue_status_plot)
```

#-----------------------------------------------
# Boxplots and Phenology

## Setup / Wrangling

### Phenology
```{r}
pheno_df_leaves <- pheno_df %>% 
  filter(category %in% c("Leaves", "Needles"), 
         phenophase_status == "Yes") %>% 
  mutate(lfm = case_when(
    phenophase_name == "Young leaves" ~ 60, 
    phenophase_name == "Young needles" ~ 55, 
    phenophase_name == "Breaking leaf buds" ~ 40,
    phenophase_name == "Breaking needle buds" ~ 35,
    phenophase_name == "Emerging needles" ~ 45, 
    phenophase_name == "Increasing leaf size" ~ 50, 
    phenophase_name == "Leaves" ~ 70, 
    phenophase_name == "Colored leaves" ~ 80
  )) %>% 
  mutate(phenophase_name = fct_relevel(phenophase_name,
                                      "Breaking needle buds",
                                       "Emerging needles",
                                       "Young needles",
                                       "Breaking leaf buds",
                                       "Increasing leaf size",
                                       "Young leaves", 
                                       "Leaves", 
                                       "Colored leaves"
  )) %>% 
  mutate(week = week(date)) 

pheno_df_all <- pheno_df %>% 
  filter(#category %in% c("Leaves", "Needles"), 
         phenophase_status == "Yes") %>% 
  mutate(line_position = case_when(
    phenophase_name == "Young leaves" ~ 60, 
    phenophase_name == "Young needles" ~ 55, 
    phenophase_name == "Breaking leaf buds" ~ 40,
    phenophase_name == "Breaking needle buds" ~ 35,
    phenophase_name == "Emerging needles" ~ 45, 
    phenophase_name == "Increasing leaf size" ~ 50, 
    phenophase_name == "Leaves" ~ 55, 
    phenophase_name == "Colored leaves" ~ 57,
    phenophase_name == "Flowers or flower buds" ~ 75, 
    phenophase_name == "Open flowers" ~ 76,
    phenophase_name == "Pollen cones" ~ 85,
    phenophase_name == "Pollen release" ~ 87, 
    phenophase_name == "Fruits" ~ 78, 
    phenophase_name == "Open pollen cones" ~ 86,
    phenophase_name == "Unripe seed cones" ~ 90,
    phenophase_name == "Recent fruit of seed drop" ~95,
    phenophase_name == "Ripe fruits" ~ 92.5,
    phenophase_name == "Ripe seed codes" ~ 91
  )) %>% 
  mutate(week = week(date)) 

unique(pheno_df_all$phenophase_name)
```

For figure with lfm boxplots: 
```{r}
pheno_fig <- pheno_df %>% 
  filter(phenophase_status == "Yes") %>% 
  mutate(lfm = case_when(
    phenophase_name == "Young leaves" ~ 20, 
    phenophase_name == "Young needles" ~ 20, 
    phenophase_name == "Breaking leaf buds" ~ 10,
    phenophase_name == "Breaking needle buds" ~ 10,
    phenophase_name == "Emerging needles" ~ 30, 
    phenophase_name == "Increasing leaf size" ~ 30, 
    phenophase_name == "Leaves" ~ 40, 
    phenophase_name == "Colored leaves" ~ 50,
    phenophase_name == "Flowers or flower buds" ~ 30,
    phenophase_name == "Pollen cones" ~ 40,
    phenophase_name == "Pollen release" ~ 50,
    phenophase_name == "Ripe fruits" ~ 40,
  TRUE ~ NA_real_
  )) %>% 
  mutate(week = week(date)) %>% 
  select(phenophase_name, lfm, date, week, spp) %>% 
  drop_na(lfm) %>% mutate(Species = case_when(
    spp == "ABCO" ~ "Ab. concolor",
    spp == "PIJE" ~ "Pi. jeffreyi", 
    spp == "CADE" ~ "Ca. decurrens", 
    spp == "CECO" ~ "Ce. cordulatus", 
    spp == "ARPA" ~ "Ar. patula", 
    spp == "QUKE" ~ "Qu. kelloggii"
    )) 
```

Phenology df data wrangling:
```{r}
pheno_fig2 <- pheno_fig %>% 
  filter(phenophase_name %in% c("Breaking needle buds", "Young needles", "Emerging needles", "Pollen cones", "Pollen release", "Breaking leaf buds", "Young leaves", "Increasing leaf size", "Flowers or flower buds", "Fruits", "Ripe fruits")) %>% 
  mutate(phenophase_name = case_when(
    phenophase_name == "Breaking needle buds" ~ "Breaking buds",
    phenophase_name == "Breaking leaf buds" ~ "Breaking buds",
    phenophase_name == "Young needles" ~ "Young leaves",
    phenophase_name == "Young leaves" ~ "Young leaves",
    phenophase_name == "Emerging needles" ~ "Increasing leaf size",
    phenophase_name == "Pollen cones" ~ "Pollen cones",
    phenophase_name == "Pollen release" ~ "Pollen release",
    phenophase_name == "Increasing leaf size" ~ "Increasing leaf size",
    phenophase_name == "Flowers or flower buds" ~ "Flowers or flower buds",
    phenophase_name == "Fruits" ~ "Fruits",
    phenophase_name == "Ripe fruits" ~ "Ripe fruits"))
```

For aspect ratio of phenophases plots:
```{r}
two <- c(1/20, 4)
three <- c(1/13, 3)
four <- c(1/10, 2.75)
```

### Boxplots
```{r}
field_data_2021_fig <- field_data_all_3 %>% 
  select(age, year, date, lfm, water_potential, Species, spp)

field_data_fig_2 <- bind_rows(field_data_2021_fig, pheno_fig) %>% 
  mutate(year = year(date), 
         doy = day(date))
field_data_fig_2

field_data_2021 <- field_data_fig_2 %>%
    filter(year == 2021,
      age != "both") %>% 
    mutate(week = week(date)) %>% 
    mutate(age_new = case_when(
      age %in% c("new") & spp %in% c("ABCO") & week %in% c(17, 20, 24) ~ "old",
      age %in% c("old") & spp %in% c("CADE") & week > 24 ~ "new",
      age %in% c("old") & spp %in% c("CECO") & week > 24 ~ "new",
      age %in% c("new") & spp %in% c("PIJE") & week %in% c(17, 20) ~ "old",
      age %in% c("old") & spp %in% c("ARPA") & week > 25 ~ "new",
      TRUE ~ as.character(age))) %>% 
    drop_na(spp, Species) %>% 
    mutate(age_labs = case_when(
      age_new %in% c("new") ~ "Current Year",
      age_new %in% c("old") ~ "Previous Year", 
    TRUE ~ as.character(age)))

spp.labs <- c("Ab. concolor","Pi. jeffreyi","Ca. decurrens","Ce. cordulatus","Ar. patula","Qu. kelloggii")
names(spp.labs) <- c("ABCO","PIJE","CADE","CECO","ARPA","QUKE")
```


## Function: Pheno x Boxplot
```{r}
pheno.boxplot.vis <- function(species, phenophases, max.y = 350, colors, no.phenophases = two, pheno_df = pheno_fig2, letter){
  
df1 <- field_data_2021 %>%
  mutate(month = case_when(
    date == '2021-04-24' ~ "April",
    date == '2021-05-14' ~ "May",
    date == '2021-05-15' ~ "May",
    date == '2021-06-15' ~ "June",
    date == '2021-06-16' ~ "June",
    date == '2021-07-08' ~ "E. July",
    date == '2021-07-09' ~ "E. July",
    date == '2021-07-30' ~ "L. July",
    date == '2021-07-31' ~ "L. July",
    date == '2021-08-19' ~ "Aug.",
    date == '2021-08-20' ~ "Aug.",
    date == '2021-09-16' ~ "Sept.",
    date == '2021-09-17' ~ "Sept.",
    date == '2021-10-16' ~ "Oct.",
    date == '2021-10-17' ~ "Oct."
  )) %>% 
  filter(age != "both",
         lfm < 350,
         spp == species)

df1$month <- factor(df1$month, levels = c("April", "May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))

df2 <- field_data_all_3 %>% 
  filter(spp == species)
mean_lfm <- mean(df2$lfm_tlp)

df3 <- pheno_df %>%
  filter(spp == species) %>% 
  filter(phenophase_name %in% phenophases)

p1 <- ggplot(data = df1, aes(x = month, y = lfm)) +
  geom_boxplot(aes(fill = age_labs)) +
  geom_hline(aes(yintercept  = mean_lfm),
             linetype = "dotted") +
  scale_y_continuous(limits = c(0, max.y)) +
  scale_fill_manual(values = c('#B2BEB5', '#36454F')) +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs)) +
  annotate('text', x = 0.875, y = max.y - 0.04*max.y, label = letter, size = 12, fontface = 'bold') +
  theme(axis.ticks.x = element_blank(),
    axis.text.x = element_text(
      angle = 30,
      hjust = 1,
      size = 16),
    axis.title = element_blank(),
    #axis.title.y = element_text(face = "bold", size = 16),
    axis.text.y = element_text(size = 12),
    legend.title = element_text(face = "bold"),
    strip.text = element_text(face = "italic", size = 14),
    legend.position = 'none') +
  labs(y = "Live Fuel Moisture",
       fill = "Tissue Status")

p2 <- ggplot2::ggplot(data = df3, aes(x = date, y = lfm, color = phenophase_name)) +
  geom_line(alpha = 0.6, size = no.phenophases[2]) +
  scale_color_manual(values = colors) +
  theme(aspect.ratio = no.phenophases[1],
        legend.position = 'none',
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill = 'white'))

p3 <- p1 + annotation_custom(ggplotGrob(p2), xmin = "April", xmax = "Oct.",
                       ymin = -10, ymax = 30)
return(p3)
  }
```

# ---

## Version 1

### QUKE
QUKE wasn't working in the function because of the lack of samples in April, so I had alter the code in the function to make it work:
```{r}
df1 <- field_data_2021 %>%
  mutate(month = case_when(
    date == '2021-05-14' ~ "May",
    date == '2021-05-15' ~ "May",
    date == '2021-06-15' ~ "June",
    date == '2021-06-16' ~ "June",
    date == '2021-07-08' ~ "E. July",
    date == '2021-07-09' ~ "E. July",
    date == '2021-07-30' ~ "L. July",
    date == '2021-07-31' ~ "L. July",
    date == '2021-08-19' ~ "Aug.",
    date == '2021-08-20' ~ "Aug.",
    date == '2021-09-16' ~ "Sept.",
    date == '2021-09-17' ~ "Sept.",
    date == '2021-10-16' ~ "Oct.",
    date == '2021-10-17' ~ "Oct."
  )) %>% 
  filter(age != "both",
         lfm < 350,
         spp == "QUKE")
df1$month <- factor(df1$month, levels = c("May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
df2 <- field_data_all_3 %>% 
  filter(spp == "QUKE")

mean_lfm <- mean(df2$lfm_tlp)

df3 <- pheno_fig2 %>%
  filter(spp == "QUKE") %>% 
  filter(phenophase_name %in% c("Breaking buds", "Increasing leaf size", "Ripe fruits"))

p1 <- ggplot(data = df1, aes(x = month, y = lfm)) +
  geom_boxplot(aes(fill = age_labs)) +
  geom_hline(aes(yintercept  = mean_lfm),
             linetype = "dotted") +
  scale_y_continuous(limits = c(0, 260)) +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs)) +
  scale_fill_manual(values = c('#B2BEB5', '#36454F')) +
 # geom_line(aes(x = month, y = lfm), color = "black") +
  annotate('text', x = 0.875, y = 249.6, label = 'f', size = 12, fontface = 'bold') +
  theme(axis.ticks.x = element_blank(),
    axis.text.x = element_text(
      angle = 30,
      hjust = 1,
      size = 16),
    #axis.title.y = element_text(face = "bold", size = 16),
    axis.text.y = element_text(size = 12),
    axis.title = element_blank(),
    legend.title = element_text(face = "bold"),
    strip.text = element_text(face = "italic", size = 14),
    legend.position = 'none') +
  labs(y = "Live Fuel Moisture",
       fill = "Tissue Status")
p1

p2 <- ggplot2::ggplot(data = df3, aes(x = date, y = lfm, color = phenophase_name)) +
  geom_line(alpha = 1, size = 3) +
  scale_color_manual(values = c("#4ECDC4", "#2B3E37", "#070093")) +
  theme(aspect.ratio = 1/13,
        legend.position = 'none',
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill = 'white'))
p2

p.quke <- p1 + annotation_custom(ggplotGrob(p2), xmin = "May", xmax = "Oct.",
                       ymin = -10, ymax = 30)
p.quke
```

### Other Species
For all other species, the function written above was able to be utilized:
```{r}
abco.phenophases <- c("Breaking buds", "Young leaves")
abco.colors <- c("#4ECDC4", "#52E764") # Based on above vector
p.abco <- pheno.boxplot.vis(species = "ABCO", phenophases = abco.phenophases,
                            colors = abco.colors, letter = 'a') +
  geom_smooth(aes(y = ))

pije.phenophases <- c("Increasing leaf size", "Young leaves")
pije.colors <- c("#2B3E37", "#52E764")
p.pije <- pheno.boxplot.vis(species = "PIJE", phenophases = pije.phenophases,
                            colors = pije.colors, letter = 'b')

cade.phenophases <- c("Pollen cones", "Pollen release")
cade.colors <- c("#6B91BD", "#E69512")
p.cade <- pheno.boxplot.vis(species = "CADE", phenophases = cade.phenophases,
                            max.y = 150, colors = cade.colors, letter = 'c')

arpa.phenophases <- c("Breaking buds", "Flowers or flower buds", "Ripe fruits", "Young leaves")
arpa.colors <- c("#4ECDC4", "#F03D79", "#070093", "#52E764")
p.arpa <-pheno.boxplot.vis(species = "ARPA", phenophases = arpa.phenophases,
                           max.y = 210, colors = arpa.colors, no.phenophases = four, letter = 'e')

ceco.phenophases <- c("Flowers or flower buds", "Ripe fruits", "Young leaves")
ceco.colors <- c("#F03D79", "#070093", "#52E764")
p.ceco <- pheno.boxplot.vis(species = "CECO", phenophases = ceco.phenophases,
                            max.y = 260, colors = ceco.colors, no.phenophases = three, letter = 'd')

```


### Legends
```{r}
# Phenophase legend:
df.leg <- pheno_fig2 %>%
  filter(phenophase_name %in% c("Breaking buds", "Young leaves", "Increasing leaf size",  "Pollen release", "Pollen cones", "Ripe fruits", "Flowers or flower buds"))
df.leg$phenophase_name <- factor(df.leg$phenophase_name, levels = c("Breaking buds", "Young leaves", "Increasing leaf size",  "Pollen release", "Pollen cones", "Ripe fruits", "Flowers or flower buds"))

legend.plot <- ggplot() +
  geom_line(aes(y = lfm, x = date, color = phenophase_name), alpha = 0.6, 
            data = df.leg, size = 4, position=position_dodge(w=1)) +
  labs(color = "Phenophase") +
  scale_color_manual(values = c("#4ECDC4", "#52E764", "#2B3E37", "#E69512", "#6B91BD", "#070093", "#F03D79")) + #Holder color palette, matched to the phenophase vector above
  theme(legend.position = 'top',
        legend.title = element_text(face = 'bold'),
        legend.key = element_rect(fill = 'white'))
legend.plot

#replace: #52E764 for #52E764, #070093 for #070093

legend <- cowplot::get_legend(legend.plot)

# Boxplot legend:
legend.plot2 <- ggplot(data = field_data_2021, aes(x = date, y = lfm)) +
  geom_boxplot(aes(fill = age_labs)) +
  scale_fill_manual(values = c('#B2BEB5', '#36454F')) +
  labs(fill = "Tissue Status") +
  theme(legend.position = 'top',
        legend.title = element_text(face = 'bold'),
        legend.key = element_rect(fill = 'white'))
legend.plot2

legend2 <- cowplot::get_legend(legend.plot2)
```

### Arrange
```{r}
# I ended up doing some odd cowplot stuff here to get the final product we wanted
p.spp <- cowplot::plot_grid(p.abco, p.pije, p.cade, p.ceco, p.arpa, p.quke, ncol = 3)
p.spp
ggsave(here("figures", "extra-figures", "phenophase.lfm.boxplot.version1.jpg"),
       height = 10, width = 15,
       plot = p.spp)

blank_plot <- ggplot(data = df1, aes(x = lfm, y = week)) +
  geom_point(color = 'white') +
  theme_void()

legends <- cowplot::plot_grid(blank_plot, legend, legend2, ncol = 1,
                              rel_heights = c(3, 1, 1))
legends
ggsave(here("figures", "main-figures", "Fig3.Legend.jpg"),
       height = 2, width = 7,
       plot = legends)

# For some strange reason, I could not get the legends and the plots to mesh together with cowplot or ggarrange... I ended up just stitching them together on Google Slides
```

# ---

## Version 2 (MAIN)
Adjusting the above figure by removing some of the phenophases in which we made note of mistakes in observations or have a low frequency across all observed individuals. This way the phenology presented represents the population observed rather than individual anomalies.

### Histograms of phenophases
I wrote this code to see which weeks had only a couple of observations to get a better sense of how the entire population acted as an average rather than single individuals
#### Our data
```{r}
abco.hist <- pheno_fig %>% 
  filter(spp == "ABCO") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_name)
abco.hist

cade.hist <- pheno_fig %>% 
  filter(spp == "CADE") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_name)
cade.hist

pije.hist <- pheno_fig %>% 
  filter(spp == "PIJE") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_name)
pije.hist

arpa.hist <- pheno_fig %>% 
  filter(spp == "ARPA") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_name)
arpa.hist

ceco.hist <- pheno_fig %>% 
  filter(spp == "CECO") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_name)
ceco.hist

quke.hist <- pheno_fig %>% 
  filter(spp == "QUKE") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_name)
quke.hist
```

#### National phenology network data
```{r}
npn_df <- npn_df %>% 
  mutate(week = week(ymd(observation_date))) %>% 
  filter(phenophase_status == 1)

npn_df %>% 
  filter(species == "concolor") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_description)

npn_df %>% 
  filter(species == "decurrens") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_description)

npn_df %>% 
  filter(species == "jeffreyi") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_description)

npn_df %>% 
  filter(species == "patula") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_description)

npn_df %>% 
  filter(species == "cordulatus") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_description)

npn_df %>% 
  filter(species == "kelloggii") %>% 
  ggplot(aes(x = week)) +
  geom_histogram() +
  facet_wrap(~phenophase_description)
```


### Data wrangling
Using above histogram, notes from phenology notebook, and a cross-comparison with the national phenology network (NPN) data to filter out certain values:
```{r}
# ABCO
pheno_fig_abco1 <- pheno_fig2 %>% 
  filter(spp == "ABCO") %>% 
  filter(phenophase_name == "Breaking buds") %>% 
  filter(week > 16 & week < 27)
pheno_fig_abco2 <- pheno_fig2 %>% 
  filter(spp == "ABCO") %>% 
  filter(phenophase_name == "Pollen cones") %>% 
  filter(week > 16)
pheno_fig_abco3 <- pheno_fig2 %>% 
  filter(spp == "ABCO") %>% 
  filter(phenophase_name == "Young leaves") %>% 
  filter(week > 16)
pheno_fig_abco <- rbind(pheno_fig_abco1, pheno_fig_abco2, pheno_fig_abco3)

# CADE
pheno_fig_cade <- pheno_fig2 %>% 
  filter(spp == "CADE")

# PIJE
pheno_fig_pije1 <- pheno_fig2 %>% 
  filter(spp == "PIJE") %>% 
  filter(phenophase_name == "Young leaves") %>% 
  filter(week > 20)
pheno_fig_pije2 <- pheno_fig2 %>% 
  filter(spp == "PIJE") %>% 
  filter(phenophase_name == "Increasing leaf size")
pheno_fig_pije <- rbind(pheno_fig_pije1, pheno_fig_pije2)

# ARPA
pheno_fig_arpa1 <- pheno_fig2 %>% 
  filter(spp == "ARPA") %>% 
  filter(phenophase_name == "Breaking buds")
pheno_fig_arpa2 <- pheno_fig2 %>% 
  filter(spp == "ARPA") %>% 
  filter(phenophase_name == "Young leaves") %>% 
  filter(week > 20)
pheno_fig_arpa3 <- pheno_fig2 %>% 
  filter(spp == "ARPA") %>% 
  filter(phenophase_name %in% c("Flowers or flower buds", "Ripe fruits"))
pheno_fig_arpa <- rbind(pheno_fig_arpa1, pheno_fig_arpa2, pheno_fig_arpa3)

# CECO
pheno_fig_ceco1 <- pheno_fig2 %>% 
  filter(spp == "CECO") %>% 
  filter(phenophase_name == "Flowers or flower buds") %>% 
  filter(week > 17)
pheno_fig_ceco2 <- pheno_fig2 %>% 
  filter(spp == "CECO") %>% 
  filter(phenophase_name %in% c("Young leaves", "Ripe fruits"))
pheno_fig_ceco <- rbind(pheno_fig_ceco1, pheno_fig_ceco2)

# QUKE
pheno_fig_quke1 <- pheno_fig2 %>% 
  filter(spp == "QUKE") %>% 
  filter(phenophase_name == "Breaking buds")
pheno_fig_quke2 <- pheno_fig2 %>% 
  filter(spp == "QUKE") %>% 
  filter(phenophase_name == "Increasing leaf size") %>% 
  filter(week < 37)
pheno_fig_quke3 <- pheno_fig2 %>% 
  filter(spp == "QUKE") %>% 
  filter(phenophase_name == "Ripe fruits") %>% 
  filter(week > 27)
pheno_fig_quke <- rbind(pheno_fig_quke1, pheno_fig_quke2, pheno_fig_quke3)

# Together
pheno_fig_main <- rbind(pheno_fig_abco, pheno_fig_cade, pheno_fig_pije, pheno_fig_arpa, pheno_fig_ceco, pheno_fig_quke)
```

### QUKE
QUKE wasn't working in the function because of the lack of samples in April. Here, all I did was re-run the phenology barplot, 
```{r}
df3 <- pheno_fig_main %>%
  filter(spp == "QUKE") %>% 
  filter(phenophase_name %in% c("Breaking buds", "Increasing leaf size", "Ripe fruits"))

p2 <- ggplot2::ggplot(data = df3, aes(x = date, y = lfm, color = phenophase_name)) +
  geom_line(alpha = 0.6, size = 3) +
  scale_color_manual(values = c("#4ECDC4", "#2B3E37", "#070093")) +
  theme(aspect.ratio = 1/13,
        legend.position = 'none',
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill = 'white'))
p2

p.quke.main <- p1 + annotation_custom(ggplotGrob(p2), xmin = "May", xmax = "Oct.",
                       ymin = -10, ymax = 30)
p.quke.main
```

### Other Species
```{r}
p.abco.main <- pheno.boxplot.vis(species = "ABCO", phenophases = abco.phenophases,
                            colors = abco.colors, pheno_df = pheno_fig_main, letter = 'a')

p.pije.main <- pheno.boxplot.vis(species = "PIJE", phenophases = pije.phenophases,
                            colors = pije.colors, pheno_df = pheno_fig_main, letter = 'b')

p.cade.main <- pheno.boxplot.vis(species = "CADE", phenophases = cade.phenophases,
                            max.y = 150, colors = cade.colors,
                            pheno_df = pheno_fig_main, letter = 'c')

p.arpa.main <-pheno.boxplot.vis(species = "ARPA", phenophases = arpa.phenophases,
                           max.y = 210, colors = arpa.colors,
                           no.phenophases = four, pheno_df = pheno_fig_main, letter = 'e')

p.ceco.main <- pheno.boxplot.vis(species = "CECO", phenophases = ceco.phenophases,
                            max.y = 260, colors = ceco.colors,
                            no.phenophases = three, pheno_df = pheno_fig_main, letter = 'd')
```

### Arrange
```{r}
# I ended up doing some odd cowplot stuff here to get the final product we wanted
y.grob1 <- textGrob("Live Fuel Moisture (%)", 
                   rot = 90,
                   gp=gpar(fontface="bold", 
                           col="black", 
                           fontsize=24))

p.spp.main <- cowplot::plot_grid(p.abco.main, p.pije.main, p.cade.main, p.ceco.main, p.arpa.main, p.quke.main, ncol = 3) #labels = c('a', 'b', 'c', 'd', 'e', 'f'), label_size = 30
p.spp.main

p.spp.w.grob <- gridExtra::grid.arrange(arrangeGrob(p.spp.main, left = y.grob1))

ggsave(here("figures", "main-figures", "Fig3.LFM.Boxplot.and.Phenophases.jpg"),
       height = 10, width = 15,
       plot = p.spp.w.grob)
# For some strange reason, I could not get the legends and the plots to mesh together with cowplot or ggarrange... I ended up just stitching them together on Google Slides
```

# ---
# Version 3: Adding Water Potential, Replacing Boxplots w/ Points


## Dotplot

### Wrangling
```{r}
df1 <- field_data_2021 %>% 
  mutate(month = case_when(
    date == '2021-04-24' ~ "April",
    date == '2021-05-14' ~ "May",
    date == '2021-05-15' ~ "May",
    date == '2021-06-15' ~ "June",
    date == '2021-06-16' ~ "June",
    date == '2021-07-08' ~ "E. July",
    date == '2021-07-09' ~ "E. July",
    date == '2021-07-30' ~ "L. July",
    date == '2021-07-31' ~ "L. July",
    date == '2021-08-19' ~ "Aug.",
    date == '2021-08-20' ~ "Aug.",
    date == '2021-09-16' ~ "Sept.",
    date == '2021-09-17' ~ "Sept.",
    date == '2021-10-16' ~ "Oct.",
    date == '2021-10-17' ~ "Oct."
  )) %>% 
  filter(age != "both",
         lfm < 350)

lfm_df <- df1 %>% 
  group_by(month, spp, age) %>% 
  summarise(sd = sd(lfm, na.rm = T), lfm = mean(lfm, na.rm = T)) %>% 
  pivot_wider(values_from = 'lfm', names_from = 'age')

old_lfm_df <- lfm_df %>% 
  select(month, spp, sd, old) %>% 
  drop_na(old) %>% 
  rename(value = old) %>% 
  mutate(name = 'old_lfm')

new_lfm_df <- lfm_df %>% 
  select(month, spp, sd, new) %>% 
  drop_na(new) %>% 
  rename(value = new) %>% 
  mutate(name = 'new_lfm')

mpa_df <- df1 %>% 
  mutate(water_potential= -1*water_potential) %>% 
  group_by(month, spp) %>%
  summarise(sd = sd(water_potential, na.rm = T), value = mean(water_potential, na.rm = T)) %>% 
  drop_na(value) %>% 
  mutate(name = 'mpa')

dots_df <- rbind(old_lfm_df, new_lfm_df, mpa_df)
dots_df$month <- factor(dots_df$month, levels = c("April", "May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
```

### Function
```{r}
pheno.dotplot.vis <- function(species, phenophases, max.y = 500, colors, no.phenophases = two, pheno_df = pheno_fig_main, letter){
  
old_lfm_df <- old_lfm_df %>%
  filter(spp == species)
old_lfm_df$month <- factor(old_lfm_df$month, levels = c("April", "May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
new_lfm_df <- new_lfm_df %>%
  filter(spp == species)
new_lfm_df$month <- factor(new_lfm_df$month, levels = c("April", "May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
mpa_df <- mpa_df %>%
  filter(spp == species)
mpa_df$month <- factor(mpa_df$month, levels = c("April", "May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
df2 <- field_data_all_3 %>% 
  filter(spp == species)
mean_lfm <- mean(df2$lfm_tlp)
mean_mpa <- mean(df2$tlp)
coef <- mean_lfm/mean_mpa
trans_coef <- -1/coef

mean_mpa_plot <- mean_mpa*coef

df3 <- pheno_df %>%
  filter(spp == species) %>% 
  filter(phenophase_name %in% phenophases)

p1 <- ggplot(data = old_lfm_df) +
  geom_point(aes(y = value, x = month), data = old_lfm_df, color = '#262E43', size = 3.5) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = old_lfm_df, color = '#262E43') +
  geom_point(aes(y = value, x = month), data = new_lfm_df, color = '#262E43', position = position_nudge(x = -0.2), shape = 1, size = 3.5) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = new_lfm_df, color = '#262E43', position = position_nudge(x = -0.2)) +
  geom_point(aes(y = value*coef, x = month), data = mpa_df, color = '#5792CC', position = position_nudge(x = 0.2), size = 3.5) +
  geom_linerange(aes(x = month, ymin = value*coef - sd*coef, ymax = value*coef + sd*coef), data = mpa_df, color = '#5792CC', position = position_nudge(x = 0.2)) +
  geom_hline(aes(yintercept  = mean_lfm),
             linetype = "dashed", color = '#262E43') +
  geom_hline(aes(yintercept  = mean_mpa_plot),
             linetype = "dashed", color = '#5792CC') +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs)) +
  annotate('text', x = 0.875, y = max.y - 0.04*max.y, label = letter, size = 12, fontface = 'bold') +
  scale_y_continuous(limits = c(0, max.y),
                     sec.axis = sec_axis(trans = ~.*trans_coef, name = 'Water Potential (-MPa)')) +
  theme(axis.ticks.x = element_blank(),
    axis.text.x = element_text(
      angle = 30,
      hjust = 1,
      size = 16),
    #axis.title.y = element_text(face = "bold", size = 16),
    axis.title = element_blank(),
    axis.text.y = element_text(size = 12),
    legend.title = element_text(face = "bold"),
    strip.text = element_text(face = "italic", size = 14),
    legend.position = 'none') +
  labs(y = "Live Fuel Moisture (%)",
       fill = "Tissue Status")

p2 <- ggplot2::ggplot(data = df3, aes(x = date, y = lfm, color = phenophase_name)) +
  geom_line(alpha = 0.6, size = no.phenophases[2]) +
  scale_color_manual(values = colors) +
  theme(aspect.ratio = no.phenophases[1],
        legend.position = 'none',
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill = 'white'))

p3 <- p1 + annotation_custom(ggplotGrob(p2), xmin = "April", xmax = "Oct.",
                       ymin = -10, ymax = 30)
return(p3)
  }
```

### QUKE
```{r}
max.y <- 300

quke_old_lfm_df <- old_lfm_df %>%
  filter(spp == 'QUKE')
quke_old_lfm_df$month <- factor(quke_old_lfm_df$month, levels = c("May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
quke_new_lfm_df <- new_lfm_df %>%
  filter(spp == 'QUKE')
quke_new_lfm_df$month <- factor(quke_new_lfm_df$month, levels = c("May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
quke_mpa_df <- mpa_df %>%
  filter(spp == 'QUKE')
quke_mpa_df$month <- factor(quke_mpa_df$month, levels = c("May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
df2 <- field_data_all_3 %>% 
  filter(spp == 'QUKE') %>% 
  mutate(tlp = -1*tlp ) 

mean_lfm <- mean(df2$lfm_tlp)
mean_mpa <- mean(df2$tlp)
coef <- mean_lfm/mean_mpa
trans_coef <- 1/coef

mean_mpa_plot <- mean_mpa*coef

df3 <- pheno_fig_main %>%
  filter(spp == 'QUKE') %>% 
  filter(phenophase_name %in% c("Breaking buds", "Increasing leaf size", "Ripe fruits"))

p1 <- ggplot(data = quke_old_lfm_df) +
  geom_point(aes(y = value, x = month), data = quke_old_lfm_df, color = '#262E43', size = 3.5) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = quke_old_lfm_df, color = '#262E43') +
  geom_point(aes(y = value, x = month), data = quke_new_lfm_df, color = '#262E43', position = position_nudge(x = -0.2), shape = 1, size = 3.5) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = quke_new_lfm_df, color = '#262E43', position = position_nudge(x = -0.2)) +
  geom_point(aes(y = value*coef, x = month), data = quke_mpa_df, color = '#5792CC', position = position_nudge(x = 0.2), size = 3.5) +
  geom_linerange(aes(x = month, ymin = value*coef - sd*coef, ymax = value*coef + sd*coef), data = quke_mpa_df, color = '#5792CC', position = position_nudge(x = 0.2)) +
  geom_hline(aes(yintercept  = mean_lfm),
             linetype = "dashed", color = '#262E43') +
  geom_hline(aes(yintercept  = mean_mpa_plot),
             linetype = "dashed", color = '#5792CC') +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs)) +
  annotate('text', x = 0.875, y = max.y - 0.04*max.y, label = 'f', size = 12, fontface = 'bold') +
  scale_y_continuous(limits = c(0, max.y), sec.axis = sec_axis(trans = ~.*trans_coef, name = 'Water Potential (-MPa)')) +
  theme(axis.ticks.x = element_blank(),
    axis.text.x = element_text(
      angle = 30,
      hjust = 1,
      size = 16),
    #axis.title.y = element_text(face = "bold", size = 16),
    axis.title = element_blank(),
    axis.text.y = element_text(size = 12),
    legend.title = element_text(face = "bold"),
    strip.text = element_text(face = "italic", size = 14),
    legend.position = 'none') +
  labs(y = "Live Fuel Moisture (%)",
       fill = "Tissue Status")
p1

p2 <- ggplot2::ggplot(data = df3, aes(x = date, y = lfm, color = phenophase_name)) +
  geom_line(alpha = 0.6, size = 3) +
  scale_color_manual(values = c("#4ECDC4", "#2B3E37", "#070093")) +
  theme(aspect.ratio = 1/13,
        legend.position = 'none',
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill = 'white'))

p.quke <- p1 + annotation_custom(ggplotGrob(p2), xmin = "May", xmax = "Oct.",
                       ymin = -10, ymax = 30)
p.quke
```

### PIJE
NOTE: The reason we don't have dry weights for L July old is due to "melted bottle" (notes in raw dataset)
```{r}

max.y <- 370

pije_old_lfm_df1 <- old_lfm_df %>%
  filter(spp == 'PIJE')
add_pije_old_lfm_df <- data.frame(month = c('L. July'), sd = c(NA), value = c(NA), spp = c('PIJE'), name = c('mpa'))
pije_old_lfm_df <- rbind(pije_old_lfm_df1, add_pije_old_lfm_df)
pije_old_lfm_df$month <- factor(pije_old_lfm_df$month, levels = c("April", "May", "June",
                                          "E. July", "L. July", "Aug.", "Sept.", "Oct."))
pije_new_lfm_df <- new_lfm_df %>%
  filter(spp == 'PIJE')
pije_new_lfm_df$month <- factor(pije_new_lfm_df$month, levels = c("April", "May", "June",
                                          "E. July", "L. July", "Aug.", "Sept.", "Oct."))
pije_mpa_df1 <- mpa_df %>%
  filter(spp == 'PIJE')
add_pije_mpa_df <- data.frame(month = c('May', 'E. July'), sd = c(NA, NA), value = c(NA, NA), spp = c('PIJE', 'PIJE'), name = c('mpa', 'mpa'))
pije_mpa_df <- rbind(pije_mpa_df1, add_pije_mpa_df)
pije_mpa_df$month <- factor(pije_mpa_df$month, levels = c("April", "May", "June",
                                          "E. July", "L. July", "Aug.", "Sept.", "Oct."))
df2 <- field_data_all_3 %>% 
  filter(spp == 'PIJE')
mean_lfm <- mean(df2$lfm_tlp)
mean_mpa <- mean(df2$tlp)
coef <- mean_lfm/mean_mpa
trans_coef <- 1/coef

mean_mpa_plot <- mean_mpa*coef

df3 <- pheno_fig_main %>%
  filter(spp == 'PIJE') %>% 
  filter(phenophase_name %in% c("Increasing leaf size", "Young leaves"))

p1 <- ggplot(data = pije_old_lfm_df) +
  geom_point(aes(y = value, x = month), data = pije_old_lfm_df, color = '#262E43', size = 3.5) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = pije_old_lfm_df, color = '#262E43') +
  geom_point(aes(y = value, x = month), data = pije_new_lfm_df, color = '#262E43', position = position_nudge(x = -0.2), shape = 1, size = 3.5) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = pije_new_lfm_df, color = '#262E43', position = position_nudge(x = -0.2)) +
  geom_point(aes(y = value*coef, x = month), data = pije_mpa_df, color = '#5792CC', position = position_nudge(x = 0.2), size = 3.5) +
  geom_linerange(aes(x = month, ymin = value*coef - sd*coef, ymax = value*coef + sd*coef), data = pije_mpa_df, color = '#5792CC', position = position_nudge(x = 0.2)) +
  geom_hline(aes(yintercept  = mean_lfm),
             linetype = "dashed", color = '#262E43') +
  geom_hline(aes(yintercept  = mean_mpa_plot),
             linetype = "dashed", color = '#5792CC') +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs)) +
  annotate('text', x = 0.875, y = max.y - 0.04*max.y, label = 'f', size = 12, fontface = 'bold') +
  scale_y_continuous(limits = c(0, max.y), sec.axis = sec_axis(trans = ~.*trans_coef, name = 'Water Potential (-MPa)')) +
  theme(axis.ticks.x = element_blank(),
    axis.text.x = element_text(
      angle = 30,
      hjust = 1,
      size = 16),
    #axis.title.y = element_text(face = "bold", size = 16),
    axis.title = element_blank(),
    axis.text.y = element_text(size = 12),
    legend.title = element_text(face = "bold"),
    strip.text = element_text(face = "italic", size = 14),
    legend.position = 'none') +
  labs(y = "Live Fuel Moisture (%)",
       fill = "Tissue Status")
p1

p2 <- ggplot2::ggplot(data = df3, aes(x = date, y = lfm, color = phenophase_name)) +
  geom_line(alpha = 0.6, size = 4) +
  scale_color_manual(values = c("#4ECDC4", "#2B3E37", "#070093")) +
  theme(aspect.ratio = 0.05,
        legend.position = 'none',
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill = 'white'))

p.pije <- p1 + annotation_custom(ggplotGrob(p2), xmin = "April", xmax = "Oct.",
                       ymin = -10, ymax = 30)
p.pije
```

### Other Species
Function did not work for QUKE or PIJE...
```{r}
abco.phenophases <- c("Breaking buds", "Young leaves")
abco.colors <- c("#4ECDC4", "#52E764") # Based on above vector
p.abco <- pheno.dotplot.vis(species = "ABCO", phenophases = abco.phenophases,
                            colors = abco.colors, letter = 'a', max.y = 470)

#pije.phenophases <- c("Increasing leaf size", "Young leaves")
#pije.colors <- c("#2B3E37", "#52E764")
#p.pije <- pheno.dotplot.vis(species = "PIJE", phenophases = pije.phenophases,
#                            colors = pije.colors, letter = 'b', max.y = 370)

cade.phenophases <- c("Pollen cones", "Pollen release")
cade.colors <- c("#6B91BD", "#E69512")
p.cade <- pheno.dotplot.vis(species = "CADE", phenophases = cade.phenophases,
                            colors = cade.colors, letter = 'c', max.y = 170)

ceco.phenophases <- c("Flowers or flower buds", "Ripe fruits", "Young leaves")
ceco.colors <- c("#F03D79", "#070093", "#52E764")
p.ceco <- pheno.dotplot.vis(species = "CECO", phenophases = ceco.phenophases,
                            colors = ceco.colors, no.phenophases = three, letter = 'd',
                            max.y = 350)

arpa.phenophases <- c("Breaking buds", "Flowers or flower buds", "Ripe fruits", "Young leaves")
arpa.colors <- c("#4ECDC4", "#F03D79", "#070093", "#52E764")
p.arpa <-pheno.dotplot.vis(species = "ARPA", phenophases = arpa.phenophases,
                           colors = arpa.colors, no.phenophases = four, letter = 'e',
                           max.y = 230)

p.arpa
```

### Dots Legend
```{r}
legend.dots.plot <- dots_df %>% 
  mutate(name = case_when(name == 'mpa' ~ 'Water Potential',
                          name == 'old_lfm' ~ 'Live Fuel Moisture \n(Old Tissue)',
                          name == 'new_lfm' ~ 'Live Fuel Moisture \n(New Tissue)')) %>% 
  ggplot(aes(x = month, y = value, color = name, shape = name)) +
    geom_point(size = 5) +
    geom_linerange(aes(ymin = value - 2*sd, ymax = value + 2*sd)) +
    scale_color_manual(values = c('#262E43', '#262E43', '#5792CC')) +
    scale_shape_manual(values = c(1, 16, 16)) +
    labs(color = "Variable", shape = "Variable") +
    theme(legend.position = 'top',
        legend.title = element_text(face = 'bold'),
        legend.key = element_rect(fill = 'white'))
legend.dots.plot

legend.dots <- cowplot::get_legend(legend.dots.plot)
```


### Arrange
```{r}
# Plots
y.grob2 <- textGrob("Water Potential (MPa)", 
                   rot = 270,
                   gp=gpar(fontface="bold", 
                           col="black", 
                           fontsize=24))

p.spp <- cowplot::plot_grid(p.abco, p.pije, p.cade, p.ceco, p.arpa, p.quke, ncol = 3)
p.spp

dots.spp.w.grob <- gridExtra::grid.arrange(arrangeGrob(p.spp, left = y.grob1, right = y.grob2))

ggsave(here("figures", "main-figures", "Fig3.Updated.LFM.Dots.and.Phenophases.jpg"),
       height = 10, width = 15,
       plot = dots.spp.w.grob)


# Legend
dots.legends <- cowplot::plot_grid(blank_plot, legend.dots, legend, ncol = 1,
                              rel_heights = c(5, .6, 1))
dots.legends
ggsave(here("figures", "main-figures", "Fig3.Updated.Dots.Legend.jpg"),
       height = 5, width = 7,
       plot = dots.legends)
```

## Checking out Water Potential Variability
```{r}
df1$month <- factor(df1$month, levels = c("April", "May", "June",
                                          "E. July", "L. July", "Aug.", "Sept.", "Oct."))

ggplot(df1, aes(x = month, y = water_potential)) +
  geom_boxplot() +
  facet_wrap(~Species, scales = 'free', ncol = 1) +
  theme_bw()

ggsave(here("figures", "extra-figures", "WaterPotential_FieldData2021.jpg"),
       height = 10, width = 5)
```


# ---
# Version 4: Adding Water Potential, Adding lines between points, replacing Boxplots w/ Points

- Adjusting position nudge to bring points closer togther
- adding geom_line so that the points for each group are connected to each other

NOTE: PIE 

## Dotplot

### Wrangling
```{r}
df1 <- field_data_2021 %>% 
  mutate(month = case_when(
    date == '2021-04-24' ~ "April",
    date == '2021-05-14' ~ "May",
    date == '2021-05-15' ~ "May",
    date == '2021-06-15' ~ "June",
    date == '2021-06-16' ~ "June",
    date == '2021-07-08' ~ "E. July",
    date == '2021-07-09' ~ "E. July",
    date == '2021-07-30' ~ "L. July",
    date == '2021-07-31' ~ "L. July",
    date == '2021-08-19' ~ "Aug.",
    date == '2021-08-20' ~ "Aug.",
    date == '2021-09-16' ~ "Sept.",
    date == '2021-09-17' ~ "Sept.",
    date == '2021-10-16' ~ "Oct.",
    date == '2021-10-17' ~ "Oct."
  )) %>% 
  filter(age != "both",
         lfm < 350) %>% 
  mutate(age = case_when(
    spp %in% c("ABCO") & month %in% c("April", "May", "June") ~ "old", 
    TRUE ~ as.character(age)
  ))

pije_lfm <- df1 %>% 
  filter(spp %in% c("PIJE"))

lfm_df <- df1 %>% 
  filter(lfm > 60) %>% 
  group_by(month, spp, age) %>% 
  summarise(sd = sd(lfm, na.rm = T), lfm = mean(lfm, na.rm = T)) %>% 
  pivot_wider(values_from = 'lfm', names_from = 'age')

old_lfm_df <- lfm_df %>% 
  select(month, spp, sd, old) %>% 
  drop_na(old) %>% 
  rename(value = old) %>% 
  mutate(name = 'old_lfm')

new_lfm_df <- lfm_df %>% 
  select(month, spp, sd, new) %>% 
  drop_na(new) %>% 
  rename(value = new) %>% 
  mutate(name = 'new_lfm')

mpa_df <- df1 %>% 
  mutate(water_potential= -1*water_potential) %>% 
  group_by(month, spp) %>%
  summarise(sd = sd(water_potential, na.rm = T), value = mean(water_potential, na.rm = T)) %>% 
  drop_na(value) %>% 
  mutate(name = 'mpa')

dots_df <- rbind(old_lfm_df, new_lfm_df, mpa_df)
dots_df$month <- factor(dots_df$month, levels = c("April", "May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
```

### Function
```{r}
pheno.dotplot.vis <- function(species, phenophases, max.y = 500, colors, no.phenophases = two, pheno_df = pheno_fig_main, letter){
  
old_lfm_df <- old_lfm_df %>%
  filter(spp == species)
old_lfm_df$month <- factor(old_lfm_df$month, levels = c("April", "May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
new_lfm_df <- new_lfm_df %>%
  filter(spp == species)
new_lfm_df$month <- factor(new_lfm_df$month, levels = c("April", "May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
mpa_df <- mpa_df %>%
  filter(spp == species)
mpa_df$month <- factor(mpa_df$month, levels = c("April", "May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
df2 <- field_data_all_3 %>% 
  filter(spp == species)
mean_lfm <- mean(df2$lfm_tlp)
mean_mpa <- mean(df2$tlp)
coef <- -1*(mean_lfm/mean_mpa)
trans_coef <- -1/coef

mean_mpa_plot <- mean_mpa*coef

df3 <- pheno_df %>%
  filter(spp == species) %>% 
  filter(phenophase_name %in% phenophases)

p1 <- ggplot(data = old_lfm_df) +
  geom_point(aes(y = value, x = month), data = old_lfm_df, color = '#262E43', size = 3.5) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = old_lfm_df, color = '#262E43', alpha = .5) +
  geom_line(aes(y = value, x = month, group = 1), data = old_lfm_df, color = '#262E43') +#add lines connecting dots
  geom_point(aes(y = value, x = month), data = new_lfm_df, color = '#262E43', position = position_nudge(x = -0.1), shape = 1, size = 3.5) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = new_lfm_df, color = '#262E43', position = position_nudge(x = -0.1), alpha = .5) +
   geom_line(aes(y = value, x = month, group = 1), data = new_lfm_df, color = '#262E43', position = position_nudge(x = -0.1)) +#add
  geom_point(aes(y = value*coef, x = month), data = mpa_df, color = '#5792CC', position = position_nudge(x = 0.1), size = 3.5) +
  geom_linerange(aes(x = month, ymin = value*coef - sd*coef, ymax = value*coef + sd*coef), data = mpa_df, color = '#5792CC', position = position_nudge(x = 0.1), alpha = .5) +
  geom_line(aes(y = value*coef, x = month, group = 1), data = mpa_df, color = '#5792CC', position = position_nudge(x = 0.1)) +#add
  geom_hline(aes(yintercept  = mean_lfm),
             linetype = "dashed", color = '#262E43') +
  geom_hline(aes(yintercept  = mean_mpa_plot),
             linetype = "dashed", color = '#5792CC') +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs)) +
  annotate('text', x = 0.875, y = max.y - 0.04*max.y, label = letter, size = 12, fontface = 'bold') +
  scale_y_continuous(limits = c(0, max.y),
                     sec.axis = sec_axis(trans = ~.*trans_coef, name = 'Water Potential (-MPa)')) +
  theme(axis.ticks.x = element_blank(),
    axis.text.x = element_text(
      angle = 30,
      hjust = 1,
      size = 16),
    #axis.title.y = element_text(face = "bold", size = 16),
    axis.title = element_blank(),
    axis.text.y = element_text(size = 12),
    legend.title = element_text(face = "bold"),
    strip.text = element_text(face = "italic", size = 14),
    legend.position = 'none') +
  labs(y = "Live Fuel Moisture (%)",
       fill = "Tissue Status")

p2 <- ggplot2::ggplot(data = df3, aes(x = date, y = lfm, color = phenophase_name)) +
  geom_line(alpha = 0.6, size = no.phenophases[2]) +
  scale_color_manual(values = colors) +
  theme(aspect.ratio = no.phenophases[1],
        legend.position = 'none',
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill = 'white'))

p3 <- p1 + annotation_custom(ggplotGrob(p2), xmin = "April", xmax = "Oct.",
                       ymin = -10, ymax = 30)
return(p3)
  }
```

### QUKE
```{r}
max.y <- 300

quke_old_lfm_df <- old_lfm_df %>%
  filter(spp == 'QUKE')
quke_old_lfm_df$month <- factor(quke_old_lfm_df$month, levels = c("May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
quke_new_lfm_df <- new_lfm_df %>%
  filter(spp == 'QUKE')
quke_new_lfm_df$month <- factor(quke_new_lfm_df$month, levels = c("May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
quke_mpa_df <- mpa_df %>%
  filter(spp == 'QUKE')
quke_mpa_df$month <- factor(quke_mpa_df$month, levels = c("May", "June", "E. July",
                                          "L. July", "Aug.", "Sept.", "Oct."))
df2 <- field_data_all_3 %>% 
  filter(spp == 'QUKE') %>% 
  mutate(tlp = -1*tlp ) 

mean_lfm <- mean(df2$lfm_tlp)
mean_mpa <- mean(df2$tlp)
coef <- mean_lfm/mean_mpa
trans_coef <- 1/coef

mean_mpa_plot <- mean_mpa*coef

df3 <- pheno_fig_main %>%
  filter(spp == 'QUKE') %>% 
  filter(phenophase_name %in% c("Breaking buds", "Increasing leaf size", "Ripe fruits"))

p1 <- ggplot(data = quke_old_lfm_df) +
  geom_point(aes(y = value, x = month), data = quke_old_lfm_df, color = '#262E43', size = 3.5) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = quke_old_lfm_df, color = '#262E43', alpha = .5) +
  geom_line(aes(x = month, y = value, group = 1), data = quke_old_lfm_df, color = '#262E43') +
  geom_point(aes(y = value, x = month), data = quke_new_lfm_df, color = '#262E43', position = position_nudge(x = 0), shape = 1, size = 3.5) +
  geom_line(aes(x = month, y = value, group = 1), data = quke_new_lfm_df, color = '#262E43', position = position_nudge(x = 0)) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = quke_new_lfm_df, color = '#262E43', position = position_nudge(x = 0), alpha = .5) +
  geom_point(aes(y = value*coef, x = month), data = quke_mpa_df, color = '#5792CC', position = position_nudge(x = 0.1), size = 3.5) +
  geom_linerange(aes(x = month, ymin = value*coef - sd*coef, ymax = value*coef + sd*coef), data = quke_mpa_df, color = '#5792CC', position = position_nudge(x = 0.1), alpha = .5) +
  geom_line(aes(x = month, y =  value*coef, group = 1), data = quke_mpa_df, color = '#5792CC', position = position_nudge(x = 0.1)) +
  geom_hline(aes(yintercept  = mean_lfm),
             linetype = "dashed", color = '#262E43') +
  geom_hline(aes(yintercept  = mean_mpa_plot),
             linetype = "dashed", color = '#5792CC') +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs)) +
  annotate('text', x = 0.875, y = max.y - 0.04*max.y, label = 'f', size = 12, fontface = 'bold') +
  scale_y_continuous(limits = c(0, max.y), sec.axis = sec_axis(trans = ~.*trans_coef, name = 'Water Potential (-MPa)')) +
  theme(axis.ticks.x = element_blank(),
    axis.text.x = element_text(
      angle = 30,
      hjust = 1,
      size = 16),
    #axis.title.y = element_text(face = "bold", size = 16),
    axis.title = element_blank(),
    axis.text.y = element_text(size = 12),
    legend.title = element_text(face = "bold"),
    strip.text = element_text(face = "italic", size = 14),
    legend.position = 'none') +
  labs(y = "Live Fuel Moisture (%)",
       fill = "Tissue Status")

p2 <- ggplot2::ggplot(data = df3, aes(x = date, y = lfm, color = phenophase_name)) +
  geom_line(alpha = 0.6, size = 3) +
  scale_color_manual(values = c("#4ECDC4", "#2B3E37", "#070093")) +
  theme(aspect.ratio = 1/13,
        legend.position = 'none',
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill = 'white'))

p.quke <- p1 + annotation_custom(ggplotGrob(p2), xmin = "May", xmax = "Oct.",
                       ymin = -10, ymax = 30)
p.quke
```

### PIJE
```{r}
max.y <- 370

pije_old_lfm_df1 <- old_lfm_df %>%
  filter(spp == 'PIJE')
add_pije_old_lfm_df <- data.frame(month = c('L. July'), sd = c(NA), value = c(NA), spp = c('PIJE'), name = c('mpa'))
pije_old_lfm_df <- rbind(pije_old_lfm_df1, add_pije_old_lfm_df) %>% 
  mutate(month = as.factor(month))


pije_old_lfm_df$month <- factor(pije_old_lfm_df$month, levels = c("April", "May", "June","E. July", "L. July", "Aug.", "Sept.", "Oct."))
unique(pije_old_lfm_df$month)

# pije_old_lfm_df <- pije_old_lfm_df %>%
#   filter(!is.na(value))

pije_new_lfm_df <- new_lfm_df %>%
  filter(spp == 'PIJE') %>% 
  mutate(value = case_when(
         month  %in% c("April", "May") ~ NA, 
         TRUE ~ as.numeric(value)))

pije_new_lfm_df$month <- factor(pije_new_lfm_df$month, levels = c("April", "May", "June","E. July", "L. July", "Aug.", "Sept.", "Oct."))

unique(pije_new_lfm_df$month)


pije_mpa_df1 <- mpa_df %>%
  filter(spp == 'PIJE')
add_pije_mpa_df <- data.frame(month = c('May', 'E. July'), sd = c(NA, NA), value = c(NA, NA), spp = c('PIJE', 'PIJE'), name = c('mpa', 'mpa'))

unique(pije_mpa_df$month)


pije_mpa_df <- rbind(pije_mpa_df1, add_pije_mpa_df)
pije_mpa_df$month <- factor(pije_mpa_df$month, levels = c("April", "May", "June", "E. July", "L. July", "Aug.", "Sept.", "Oct."))

pije_mpa_df <- pije_mpa_df %>% 
  filter(!is.na(value)) 

df2 <- field_data_all_3 %>% 
  filter(spp == 'PIJE')
mean_lfm <- mean(df2$lfm_tlp)
mean_mpa <- mean(df2$tlp)
coef <- -1*(mean_lfm/mean_mpa)
trans_coef <- 1/coef
mean_mpa_plot <- mean_mpa*coef

df3 <- pheno_fig_main %>%
  filter(spp == 'PIJE') %>% 
  filter(phenophase_name %in% c("Increasing leaf size", "Young leaves"))

pije_old_lfm_df_filtered <- pije_old_lfm_df[complete.cases(pije_old_lfm_df), ]

p1 <- ggplot(data = pije_old_lfm_df) +
  geom_point(aes(y = value, x = month), data = pije_new_lfm_df, color = '#262E43', position = position_nudge(x = -0.1), shape = 1, size = 3.5) +
  geom_line(aes(y = value, x = month, group = 1
                ), data = pije_new_lfm_df, color = '#262E43', position = position_nudge(x = -0.1)
            ) +#add lines connecting dots
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = pije_new_lfm_df, color = '#262E43', position = position_nudge(x = -0.1), alpha = .5) +
  geom_point(aes(y = value, x = month), data = pije_old_lfm_df, color = '#262E43', size = 3.5) +
  geom_linerange(aes(x = month, ymin = value - sd, ymax = value + sd), data = pije_old_lfm_df, color = '#262E43', alpha = .5) +
  geom_line(aes(y = value, x = month, group = 1), data =pije_old_lfm_df_filtered, color = '#262E43', na.rm = TRUE) +#add lines connecting dots
  geom_line(aes(y = value*coef, x = month), data = pije_mpa_df, color = '#5792CC', 
            position = position_nudge(x = 0.1)) +
  geom_point(aes(y = value*coef, x = month), data = pije_mpa_df, color = '#5792CC', position = position_nudge(x = 0.1), size = 3.5) +
  geom_linerange(aes(x = month, ymin = value*coef - sd*coef, ymax = value*coef + sd*coef), data = pije_mpa_df, color = '#5792CC', position = position_nudge(x = 0.1), alpha = .5) +
  geom_line(aes(y = value*coef, x = month, group = 2), data = pije_mpa_df, color = '#5792CC', position = position_nudge(x = 0.1)) + #add lines connecting dots
  geom_hline(aes(yintercept  = mean_lfm),
             linetype = "dashed", color = '#262E43') +
  geom_hline(aes(yintercept  = mean_mpa_plot),
             linetype = "dashed", color = '#5792CC') +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs)) +
  annotate('text', x = 0.875, y = max.y - 0.04*max.y, label = 'f', size = 12, fontface = 'bold') +
  scale_y_continuous(limits = c(0, max.y), sec.axis = sec_axis(trans = ~.*trans_coef, name = 'Water Potential (-MPa)')) +
  theme(axis.ticks.x = element_blank(),
    axis.text.x = element_text(
      angle = 30,
      hjust = 1,
      size = 16),
    #axis.title.y = element_text(face = "bold", size = 16),
    axis.title = element_blank(),
    axis.text.y = element_text(size = 12),
    legend.title = element_text(face = "bold"),
    strip.text = element_text(face = "italic", size = 14),
    legend.position = 'none') +
  labs(y = "Live Fuel Moisture (%)",
       fill = "Tissue Status")
p1

p2 <- ggplot2::ggplot(data = df3, aes(x = date, y = lfm, color = phenophase_name)) +
  geom_line(alpha = 0.6, size = 4) +
  scale_color_manual(values = c("#4ECDC4", "#2B3E37", "#070093")) +
  theme(aspect.ratio = 0.05,
        legend.position = 'none',
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        panel.background = element_rect(fill = 'white'))

p.pije <- p1 + annotation_custom(ggplotGrob(p2), xmin = "April", xmax = "Oct.",
                       ymin = -10, ymax = 30)
p.pije
```


### Other Species
Function did not work for QUKE or PIJE...
```{r}
abco.phenophases <- c("Breaking buds", "Young leaves")
abco.colors <- c("#4ECDC4", "#52E764") # Based on above vector
p.abco <- pheno.dotplot.vis(species = "ABCO", phenophases = abco.phenophases,
                            colors = abco.colors, letter = 'a', max.y = 470)
p.abco

#pije.phenophases <- c("Increasing leaf size", "Young leaves")
#pije.colors <- c("#2B3E37", "#52E764")
#p.pije <- pheno.dotplot.vis(species = "PIJE", phenophases = pije.phenophases,
#                            colors = pije.colors, letter = 'b', max.y = 370)

cade.phenophases <- c("Pollen cones", "Pollen release")
cade.colors <- c("#6B91BD", "#E69512")
p.cade <- pheno.dotplot.vis(species = "CADE", phenophases = cade.phenophases,
                            colors = cade.colors, letter = 'c', max.y = 170)
p.cade

ceco.phenophases <- c("Flowers or flower buds", "Ripe fruits", "Young leaves")
ceco.colors <- c("#F03D79", "#070093", "#52E764")
p.ceco <- pheno.dotplot.vis(species = "CECO", phenophases = ceco.phenophases,
                            colors = ceco.colors, no.phenophases = three, letter = 'd',
                            max.y = 350)
p.ceco

arpa.phenophases <- c("Breaking buds", "Flowers or flower buds", "Ripe fruits", "Young leaves")
arpa.colors <- c("#4ECDC4", "#F03D79", "#070093", "#52E764")
p.arpa <-pheno.dotplot.vis(species = "ARPA", phenophases = arpa.phenophases,
                           colors = arpa.colors, no.phenophases = four, letter = 'e',
                           max.y = 230)
p.arpa
```

### Dots Legend
```{r}
legend.dots.plot <- dots_df %>% 
  mutate(name = case_when(name == 'mpa' ~ 'Water Potential',
                          name == 'old_lfm' ~ 'Live Fuel Moisture \n(Old/Developed Tissue)',
                          name == 'new_lfm' ~ 'Live Fuel Moisture \n(New Tissue)')) %>% 
  ggplot(aes(x = month, y = value, color = name, shape = name)) +
    geom_point(size = 5) +
    geom_linerange(aes(ymin = value - 2*sd, ymax = value + 2*sd)) +
    scale_color_manual(values = c('#262E43', '#262E43', '#5792CC')) +
    scale_shape_manual(values = c(1, 16, 16)) +
    labs(color = "Variable", shape = "Variable") +
    theme(legend.position = 'top',
        legend.title = element_text(face = 'bold'),
        legend.key = element_rect(fill = 'white'))
legend.dots.plot

legend.dots <- cowplot::get_legend(legend.dots.plot)
```


### Arrange
```{r}
# Plots
y.grob2 <- textGrob("Water Potential (MPa)", 
                   rot = 270,
                   gp=gpar(fontface="bold", 
                           col="black", 
                           fontsize=24))

p.spp <- cowplot::plot_grid(p.abco, p.pije, p.cade, p.ceco, p.arpa, p.quke, ncol = 3)
p.spp

dots.spp.w.grob <- gridExtra::grid.arrange(arrangeGrob(p.spp, left = y.grob1, right = y.grob2))

ggsave(here("figures", "main-figures", "Fig3.Updated.LFM.Dots.and.Phenophases.jpg"),
       height = 10, width = 15,
       plot = dots.spp.w.grob)


# Legend
dots.legends <- cowplot::plot_grid(blank_plot, legend.dots, legend, ncol = 1,
                              rel_heights = c(5, .6, 1))
dots.legends

showtext::showtext_auto()
showtext::showtext_opts(dpi = 300)
ggsave(here("figures", "main-figures", "Fig3.Updated.Dots.Legend.jpg"),
       height = 5, width = 7,
       plot = dots.legends)
```

## Checking out Water Potential Variability
```{r}
df1$month <- factor(df1$month, levels = c("April", "May", "June",
                                          "E. July", "L. July", "Aug.", "Sept.", "Oct."))

ggplot(df1, aes(x = month, y = water_potential)) +
  geom_boxplot() +
  facet_wrap(~Species, scales = 'free', ncol = 1) +
  theme_bw()

ggsave(here("figures", "extra-figures", "WaterPotential_FieldData2021.jpg"),
       height = 10, width = 5)
```


#-----------------------------------------------
# Hydroscapes
Based on Hartmann 2020

```{r}
p1 <- field_data_all_3 %>% 
 filter(age == "new", strategy == "Isohydric") %>% #filtering to 'old' so that we dont have duplicate MPa datapoints
  #filter(site == 1) %>% 

   mutate(Species = fct_relevel(spp,
                "Ab. concolor",
    "Pi. jeffreyi", 
    "Ca. decurrens")) %>% 
  ggplot(aes(y = midday, x = predawn,
            # shape = as.factor(site)
             )) +
  geom_point(size = 1, 
             alpha = .5
             #aes(color = doy)
             ) +
  facet_wrap(~Species) +
 #annotate("text", x= -1, y= -5, aes(label=strategy), size = 3) +
  geom_abline(size = 1) +
  geom_smooth(method = "lm", 
              color = "#7c668c", 
              size = 1, se = FALSE)+
  ggpubr::stat_regline_equation(size = 4) +
  geom_hline(aes(yintercept = tlp), size = 1, 
             color = "#db8872", 
             linetype = "dotted")+
  
  labs(y= "", 
      x = "", 
      # color = "Day of Year",
       title = "More Isohydric"
      ) +
  theme(legend.position = "none", plot.margin = unit(c(0,0,0,0), "cm"))  + 
  scale_x_continuous(labels = NULL, breaks = NULL) +
  coord_equal()
p1

p2 <- field_data_all_3 %>% 
 filter(age == "new", strategy == "Anisohydric") %>% #filtering to 'old' so that we dont have duplicate MPa datapoints
  #filter(site == 1) %>% 
   mutate(Species = fct_relevel(spp,
        "Ar. patula", 
        "Ce. cordulatus", 
        "Qu. kelloggii")) %>% 
  ggplot(aes(y = midday, x = predawn,
            # shape = as.factor(site)
             )) +
  geom_point(size = 1, 
             alpha = .5
             #aes(color = doy)
             ) +
  facet_wrap(~Species) +
 #annotate("text", x= -1, y= -5, aes(label=strategy), size = 3) +
  geom_abline(size = 1) +
  geom_smooth(method = "lm", color = "#7c668c", size = 1, se = FALSE) +
  ggpubr::stat_regline_equation(size = 4) +
  geom_hline(aes(yintercept = tlp), 
             linewidth = 1, 
             color = "#db8872", 
             linetype = "dotted")+
  labs(y= "", 
       x = "", 
      # color = "Day of Year", 
       title = "More Anisohydric"
      ) +
  theme(legend.position = "none", 
        plot.margin = unit(c(0,0,0,0), "cm"), 
        plot.title = element_text(hjust = 1)) +
  coord_equal()
p2

plot <- cowplot::plot_grid(p1, p2,
          nrow = 2,
  label_size = 12,
  align = "hv"
  
  #labels = c("Isohydric", "Anisohydric")  
)

#create common x and y labels
y.grob <- textGrob("Midday Water Potential (MPa)", 
                   gp=gpar(col="black", fontsize=15), rot=90)

x.grob <- textGrob("Predawn Water Potential (MPa)", 
                   gp=gpar(col="black", fontsize=15))

#add to plot
grid.arrange(arrangeGrob(plot, 
                         left = y.grob, 
                         bottom = x.grob))

```

## July - October
```{r}
p1 <- field_data_all_3 %>% 
 filter(age == "new",
        strategy == "Isohydric",
        month > 6
        ) %>% 
  ggplot(aes(y = midday, x = predawn,
            # shape = as.factor(site)
             )) +
  geom_point(size = 1, 
             alpha = .5
             #aes(color = doy)
             ) +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs)) +
 #annotate("text", x= -1, y= -5, aes(label=strategy), size = 3) +
  geom_abline(size = 1) +
  geom_smooth(method = "lm", 
              color = "#7c668c", 
              size = 1, se = FALSE)+
  ggpubr::stat_regline_equation(size = 4) +
  geom_hline(aes(yintercept = tlp), size = 1, 
             color = "#db8872", 
             linetype = "dotted")+
  
  labs(y= "", 
      x = "", 
      # color = "Day of Year",
      # title = "More Isohydric"
      ) +
  theme(legend.position = "none", plot.margin = unit(c(0,0,0,0), "cm"),
        strip.text = element_text(face = "italic", size = 12))  + 
  scale_x_continuous(labels = NULL, breaks = NULL) 
p1

p2 <- field_data_all_3 %>% 
 filter(age == "new", 
        strategy == "Anisohydric", 
         month > 6
        ) %>% #filtering to 'old' so that we dont have duplicate MPa datapoints
  #filter(site == 1) %>% 
  ggplot(aes(y = midday, x = predawn,
            # shape = as.factor(site)
             )) +
  geom_point(size = 1, 
             alpha = .5
             #aes(color = doy)
             ) +
  facet_wrap(~spp, labeller = labeller(spp = spp.labs)) +
 #annotate("text", x= -1, y= -5, aes(label=strategy), size = 3) +
  geom_abline(size = 1) +
  geom_smooth(method = "lm", color = "#7c668c", size = 1, se = FALSE) +
  ggpubr::stat_regline_equation(size = 4) +
  geom_hline(aes(yintercept = tlp), 
             size = 1, 
             color = "#db8872", 
             linetype = "dotted")+
  
  labs(y= "", 
       x = "", 
      # color = "Day of Year", 
       # title = "More Anisohydric"
      ) +
  theme(legend.position = "none", 
        plot.margin = unit(c(0,0,0,0), "cm"), 
        plot.title = element_text(hjust = 1),
        strip.text = element_text(face = "italic", size = 12))
p2

plot <- cowplot::plot_grid(p1, p2,
          nrow = 2,
  label_size = 12,
  align = "hv"
  
  #labels = c("Isohydric", "Anisohydric")  
)

#create common x and y labels
y.grob <- textGrob("Midday Water Potential (MPa)", 
                   gp=gpar(col="black", fontsize=15), rot=90)

x.grob <- textGrob("Predawn Water Potential (MPa)", 
                   gp=gpar(col="black", fontsize=15, face = "bold"))

#add to plot
grid.arrange(arrangeGrob(plot, 
                         left = y.grob, 
                         bottom = x.grob))

```

## Old vs. New
```{r}
field_data_all_3 %>%
  ggplot(aes(
    x = midday,
    y = lfm,
    color = month,
    shape = spp
  )) +
  geom_point(size = 1) +
  geom_vline(
    aes(xintercept = tlp),
    size = .5,
    color = "olivedrab",
    linetype = "dotted"
  ) +
  xlim(-5, 0) +
  labs(
    y = "Live Fuel Moisture",
    x = "Midday (-MPa)",
    color = "Month",
    shape = "Species"
  ) +
  ylim(50, 300) +
  facet_wrap(~ interaction(year, age))
```

#-----------------------------------------------
# LFM vs. MPa

## Species, Time
```{r}
field_data_all_3 <- field_data_all_3 %>% 
  mutate(Age = case_when(
    age == 'new' ~ 'New',
    age == 'old' ~ 'Old'
  ))

field_data_all_3 %>%
  #filter(month < 10) %>% #removing post-rain values (October sample collection)
  mutate(spp = fct_relevel("ABCO", "PIJE", "CADE", "CECO", "ARPA", "QUKE"))  %>%
  mutate(month = month(date, label = TRUE, abbr = TRUE)) %>%
  ggplot(aes(
    x = midday,
    y = lfm,
    shape = Age
  )) +
  geom_point(size = 1.4, alpha = .65, aes(color = as.factor(month))) +
 # geom_smooth(se = F, method = 'lm', size = 0.5, alpha = .6, color = 'black', aes(lty = Age), show_guide = F) +
  geom_vline(
    aes(xintercept = tlp, #lty = Species
        ),
    size = 0.7,
    color = 'black',
    alpha = .6,
    show_guide = F
  ) +
  # geom_hline(
  #   aes(yintercept = mean_lfm_fall_spp, lty = Species),
  #   size = 0.7,
  #   color = 'black',
  #   alpha = .6,
  #   show_guide = F
  # ) +
  #stat_smooth(aes(x = LFM, y = predawn, colour = Date, linetype = "Linear Fit"), method = "lm")
  xlim(-5, 0) +
  labs(
    y = "Live Fuel Moisture",
    x = "Midday Water Potential (MPa)",
    color = "Month"
  ) +
  ylim(50, 300) +
  facet_wrap( ~ spp, labeller = labeller(spp = spp.labs)) +
  theme(
    strip.text = element_text(face = "italic"),
    axis.title = element_text(face = "bold"),
    legend.title = element_text(face = "bold"),
    strip.background = element_rect(fill = "gray90")
  ) 
  #scale_color_gradientn(colors=met.brewer("Morgenstern"))
```


## Time, Age
```{r}
field_data_all_3 %>%
   mutate(spp = fct_relevel(spp, "ABCO","PIJE","CADE","CECO", "ARPA","QUKE"))  %>% 
   mutate(Species = fct_relevel(Species,
                              "Ab. concolor",
                              "Pi. jeffreyi",
                              "Ca. decurrens",
                              "Ce. cordulatus",
                              "Ar. patula", 
                              "Qu. kelloggii"
    )) %>% 
  filter(age != "both") %>% 
  mutate(month = month(date, label = TRUE, abbr = TRUE)) %>% 
  ggplot(aes(x = predawn, 
             y = lfm, 
             color = as.factor(month), 
             shape = age)) +
  geom_point(size = 1, alpha = .7) +
  geom_vline(aes(xintercept = tlp), size = .5, alpha = .6,
             color = "black", linetype = "dotted") +
  #stat_smooth(aes(x = LFM, y = predawn, colour = Date, linetype = "Linear Fit"), method = "lm") 
  xlim(-5,0) + 
  labs( y = "Live Fuel Moisture", 
        x = "Predawn (-MPa)", 
        color = "Month", 
        shape = "Age") +
  ylim(50,300) +
  facet_wrap(~strategy)
```

## Age
```{r}
field_data_all_3 %>%
   mutate(spp = fct_relevel(spp, "ABCO","PIJE","CADE","CECO", "ARPA","QUKE"))  %>% 
   mutate(Species = fct_relevel(Species,
                              "Ab. concolor",
                              "Pi. jeffreyi",
                              "Ca. decurrens",
                              "Ce. cordulatus",
                              "Ar. patula", 
                              "Qu. kelloggii"
    )) %>% 
  filter(age != "both") %>% 
  mutate(month = month(date, label = TRUE, abbr = TRUE)) %>% 
  ggplot(aes(x = midday, 
             y = lfm, 
             #color = as.factor(month), 
             color = age)) +
  geom_point(size = 1, alpha = .7) +
  geom_vline(aes(xintercept = mean(tlp)), size = .5, alpha = .6,
             color = "black", linetype = "dotted") +
  #stat_smooth(aes(x = LFM, y = predawn, colour = Date, linetype = "Linear Fit"), method = "lm") 
  xlim(-5,0) + 
  labs( y = "Live Fuel Moisture", 
        x = "Midday (-MPa)", 
        color = "Age", 
        shape = "") +
  ylim(50,300) +
  facet_wrap(~strategy) 
```

## F. Group (Tree vs. Shrub)
```{r}
field_data_all_3 %>%
  ggplot(aes(x = midday, y = lfm, color = month, shape = spp)) +
  geom_point(size = 1) +
  geom_vline(aes(xintercept = tlp), size = .5, 
             color = "olivedrab", 
             linetype = "dotted"
             ) +
  #stat_smooth(aes(x = LFM, y = predawn, colour = Date, linetype = "Linear Fit"), method = "lm") 
  xlim(-5,0) + 
  labs( y = "Live Fuel Moisture", 
        x = "Midday (-MPa)", 
        color = "Month", 
        shape = "Species") +
  ylim(50,300) +
  facet_wrap(~type)+
  #ggtitle("Outliers removed (outlierKD)") +
  #scale_color_gradient(low = "darkblue", high = "goldenrod") +
  scale_color_gradient(low = "maroon", high = "goldenrod") 
```
